--- frontend/src/container/InfraMonitoringK8s/Clusters/ClusterDetails/Metrics/index.ts
@@ -1,3 +1,0 @@
-import ClusterMetrics from './ClusterMetrics';
-
-export default ClusterMetrics;
