+++ frontend/src/container/InfraMonitoringK8s/Clusters/ClusterDetails/constants.ts
@@ -1,5 +1,5 @@
 /* eslint-disable sonarjs/no-duplicate-string */
-import { K8sNamespacesData } from 'api/infraMonitoring/getK8sNamespacesList';
+import { K8sClustersData } from 'api/infraMonitoring/getK8sClustersList';
 import { PANEL_TYPES } from 'constants/queryBuilder';
 import { GetQueryResultsProps } from 'lib/dashboard/getQueryResults';
 import { DataTypes } from 'types/api/queryBuilder/queryAutocompleteResponse';
@@ -7,50 +7,43 @@ import { EQueryType } from 'types/common/dashboard';
 import { DataSource } from 'types/common/queryBuilder';
 import { v4 } from 'uuid';
 
-export const namespaceWidgetInfo = [
+export const clusterWidgetInfo = [
 	{
-		title: 'CPU Usage (cores)',
+		title: 'CPU Usage, allocatable',
 		yAxisUnit: '',
 	},
 	{
-		title: 'Memory Usage (bytes)',
+		title: 'Memory Usage, allocatable',
 		yAxisUnit: 'bytes',
 	},
 	{
-		title: 'Pods CPU (top 10)',
+		title: 'Ready Nodes',
 		yAxisUnit: '',
 	},
 	{
-		title: 'Pods Memory (top 10)',
-		yAxisUnit: 'bytes',
-	},
-	{
-		title: 'Network rate',
-		yAxisUnit: 'binBps',
-	},
-	{
-		title: 'Network errors',
+		title: 'NotReady Nodes',
 		yAxisUnit: '',
 	},
 	{
-		title: 'StatefulSets',
-	},
-	{
-		title: 'ReplicaSets',
+		title: 'Deployments available and desired',
 		yAxisUnit: '',
 	},
 	{
-		title: 'DaemonSets',
+		title: 'Statefulset pods',
 		yAxisUnit: '',
 	},
 	{
-		title: 'Deployments',
+		title: 'Daemonset nodes',
+		yAxisUnit: '',
+	},
+	{
+		title: 'Jobs',
 		yAxisUnit: '',
 	},
 ];
 
-export const getNamespaceQueryPayload = (
-	namespace: K8sNamespacesData,
+export const getClusterMetricsQueryPayload = (
+	cluster: K8sClustersData,
 	start: number,
 	end: number,
 ): GetQueryResultsProps[] => [
@@ -76,17 +69,17 @@ export const getNamespaceQueryPayload = (
 						filters: {
 							items: [
 								{
-									id: '47b3adae',
+									id: '0224c582',
 									key: {
 										dataType: DataTypes.String,
-										id: 'k8s_namespace_name--string--tag--false',
+										id: 'k8s_cluster_name--string--tag--false',
 										isColumn: false,
 										isJSON: false,
-										key: 'k8s_namespace_name',
+										key: 'k8s_cluster_name',
 										type: 'tag',
 									},
 									op: '=',
-									value: namespace.namespaceName,
+									value: cluster.meta.k8s_cluster_name,
 								},
 							],
 							op: 'AND',
@@ -106,49 +99,6 @@ export const getNamespaceQueryPayload = (
 					{
 						aggregateAttribute: {
 							dataType: DataTypes.Float64,
-							id: 'k8s_container_cpu_request--float64--Gauge--true',
-							isColumn: true,
-							isJSON: false,
-							key: 'k8s_container_cpu_request',
-							type: 'Gauge',
-						},
-						aggregateOperator: 'latest',
-						dataSource: DataSource.METRICS,
-						disabled: false,
-						expression: 'B',
-						filters: {
-							items: [
-								{
-									id: '93d2be5e',
-									key: {
-										dataType: DataTypes.String,
-										id: 'k8s_namespace_name--string--tag--false',
-										isColumn: false,
-										isJSON: false,
-										key: 'k8s_namespace_name',
-										type: 'tag',
-									},
-									op: '=',
-									value: namespace.namespaceName,
-								},
-							],
-							op: 'AND',
-						},
-						functions: [],
-						groupBy: [],
-						having: [],
-						legend: 'requests',
-						limit: null,
-						orderBy: [],
-						queryName: 'B',
-						reduceTo: 'avg',
-						spaceAggregation: 'sum',
-						stepInterval: 60,
-						timeAggregation: 'latest',
-					},
-					{
-						aggregateAttribute: {
-							dataType: DataTypes.Float64,
 							id: 'k8s_pod_cpu_utilization--float64--Gauge--true',
 							isColumn: true,
 							isJSON: false,
@@ -158,21 +108,21 @@ export const getNamespaceQueryPayload = (
 						aggregateOperator: 'min',
 						dataSource: DataSource.METRICS,
 						disabled: false,
-						expression: 'C',
+						expression: 'B',
 						filters: {
 							items: [
 								{
-									id: '795eb679',
+									id: 'a0e11f0c',
 									key: {
 										dataType: DataTypes.String,
-										id: 'k8s_namespace_name--string--tag--false',
+										id: 'k8s_cluster_name--string--tag--false',
 										isColumn: false,
 										isJSON: false,
-										key: 'k8s_namespace_name',
+										key: 'k8s_cluster_name',
 										type: 'tag',
 									},
 									op: '=',
-									value: namespace.namespaceName,
+									value: cluster.meta.k8s_cluster_name,
 								},
 							],
 							op: 'AND',
@@ -183,7 +133,7 @@ export const getNamespaceQueryPayload = (
 						legend: 'usage (min)',
 						limit: null,
 						orderBy: [],
-						queryName: 'C',
+						queryName: 'B',
 						reduceTo: 'avg',
 						spaceAggregation: 'sum',
 						stepInterval: 60,
@@ -201,21 +151,21 @@ export const getNamespaceQueryPayload = (
 						aggregateOperator: 'max',
 						dataSource: DataSource.METRICS,
 						disabled: false,
-						expression: 'D',
+						expression: 'C',
 						filters: {
 							items: [
 								{
-									id: '6792adbe',
+									id: 'c775629c',
 									key: {
 										dataType: DataTypes.String,
-										id: 'k8s_namespace_name--string--tag--false',
+										id: 'k8s_cluster_name--string--tag--false',
 										isColumn: false,
 										isJSON: false,
-										key: 'k8s_namespace_name',
+										key: 'k8s_cluster_name',
 										type: 'tag',
 									},
 									op: '=',
-									value: namespace.namespaceName,
+									value: cluster.meta.k8s_cluster_name,
 								},
 							],
 							op: 'AND',
@@ -226,11 +176,54 @@ export const getNamespaceQueryPayload = (
 						legend: 'usage (max)',
 						limit: null,
 						orderBy: [],
-						queryName: 'D',
+						queryName: 'C',
 						reduceTo: 'avg',
 						spaceAggregation: 'sum',
 						stepInterval: 60,
 						timeAggregation: 'max',
+					},
+					{
+						aggregateAttribute: {
+							dataType: DataTypes.Float64,
+							id: 'k8s_node_allocatable_cpu--float64--Gauge--true',
+							isColumn: true,
+							isJSON: false,
+							key: 'k8s_node_allocatable_cpu',
+							type: 'Gauge',
+						},
+						aggregateOperator: 'avg',
+						dataSource: DataSource.METRICS,
+						disabled: false,
+						expression: 'D',
+						filters: {
+							items: [
+								{
+									id: '31f6c43a',
+									key: {
+										dataType: DataTypes.String,
+										id: 'k8s_cluster_name--string--tag--false',
+										isColumn: false,
+										isJSON: false,
+										key: 'k8s_cluster_name',
+										type: 'tag',
+									},
+									op: '=',
+									value: cluster.meta.k8s_cluster_name,
+								},
+							],
+							op: 'AND',
+						},
+						functions: [],
+						groupBy: [],
+						having: [],
+						legend: 'allocatable',
+						limit: null,
+						orderBy: [],
+						queryName: 'D',
+						reduceTo: 'avg',
+						spaceAggregation: 'sum',
+						stepInterval: 60,
+						timeAggregation: 'avg',
 					},
 				],
 				queryFormulas: [],
@@ -281,17 +274,17 @@ export const getNamespaceQueryPayload = (
 						filters: {
 							items: [
 								{
-									id: '10011298',
+									id: '61a3c55e',
 									key: {
 										dataType: DataTypes.String,
-										id: 'k8s_namespace_name--string--tag--false',
+										id: 'k8s_cluster_name--string--tag--false',
 										isColumn: false,
 										isJSON: false,
-										key: 'k8s_namespace_name',
+										key: 'k8s_cluster_name',
 										type: 'tag',
 									},
 									op: '=',
-									value: namespace.namespaceName,
+									value: cluster.meta.k8s_cluster_name,
 								},
 							],
 							op: 'AND',
@@ -311,135 +304,6 @@ export const getNamespaceQueryPayload = (
 					{
 						aggregateAttribute: {
 							dataType: DataTypes.Float64,
-							id: 'k8s_container_memory_request--float64--Gauge--true',
-							isColumn: true,
-							isJSON: false,
-							key: 'k8s_container_memory_request',
-							type: 'Gauge',
-						},
-						aggregateOperator: 'latest',
-						dataSource: DataSource.METRICS,
-						disabled: false,
-						expression: 'B',
-						filters: {
-							items: [
-								{
-									id: 'ea53b656',
-									key: {
-										dataType: DataTypes.String,
-										id: 'k8s_namespace_name--string--tag--false',
-										isColumn: false,
-										isJSON: false,
-										key: 'k8s_namespace_name',
-										type: 'tag',
-									},
-									op: '=',
-									value: namespace.namespaceName,
-								},
-							],
-							op: 'AND',
-						},
-						functions: [],
-						groupBy: [],
-						having: [],
-						legend: 'request',
-						limit: null,
-						orderBy: [],
-						queryName: 'B',
-						reduceTo: 'avg',
-						spaceAggregation: 'sum',
-						stepInterval: 60,
-						timeAggregation: 'latest',
-					},
-					{
-						aggregateAttribute: {
-							dataType: DataTypes.Float64,
-							id: 'k8s_pod_memory_working_set--float64--Gauge--true',
-							isColumn: true,
-							isJSON: false,
-							key: 'k8s_pod_memory_working_set',
-							type: 'Gauge',
-						},
-						aggregateOperator: 'avg',
-						dataSource: DataSource.METRICS,
-						disabled: false,
-						expression: 'C',
-						filters: {
-							items: [
-								{
-									id: '674ace83',
-									key: {
-										dataType: DataTypes.String,
-										id: 'k8s_namespace_name--string--tag--false',
-										isColumn: false,
-										isJSON: false,
-										key: 'k8s_namespace_name',
-										type: 'tag',
-									},
-									op: '=',
-									value: namespace.namespaceName,
-								},
-							],
-							op: 'AND',
-						},
-						functions: [],
-						groupBy: [],
-						having: [],
-						legend: 'working set',
-						limit: null,
-						orderBy: [],
-						queryName: 'C',
-						reduceTo: 'avg',
-						spaceAggregation: 'sum',
-						stepInterval: 60,
-						timeAggregation: 'avg',
-					},
-					{
-						aggregateAttribute: {
-							dataType: DataTypes.Float64,
-							id: 'k8s_pod_memory_rss--float64--Gauge--true',
-							isColumn: true,
-							isJSON: false,
-							key: 'k8s_pod_memory_rss',
-							type: 'Gauge',
-						},
-						aggregateOperator: 'avg',
-						dataSource: DataSource.METRICS,
-						disabled: false,
-						expression: 'D',
-						filters: {
-							items: [
-								{
-									id: '187dbdb3',
-									key: {
-										dataType: DataTypes.String,
-										id: 'k8s_namespace_name--string--tag--false',
-										isColumn: false,
-										isJSON: false,
-										key: 'k8s_namespace_name',
-										type: 'tag',
-									},
-									op: '=',
-									value: namespace.namespaceName,
-								},
-							],
-							op: 'AND',
-						},
-						functions: [],
-						groupBy: [],
-						having: [],
-						legend: 'rss',
-						limit: null,
-						orderBy: [],
-						queryName: 'D',
-						reduceTo: 'avg',
-						spaceAggregation: 'sum',
-						stepInterval: 60,
-						timeAggregation: 'avg',
-					},
-					{
-						aggregateAttribute: {
-							dataType: DataTypes.Float64,
 							id: 'k8s_pod_memory_usage--float64--Gauge--true',
 							isColumn: true,
 							isJSON: false,
@@ -449,21 +313,21 @@ export const getNamespaceQueryPayload = (
 						aggregateOperator: 'min',
 						dataSource: DataSource.METRICS,
 						disabled: false,
-						expression: 'E',
+						expression: 'B',
 						filters: {
 							items: [
 								{
-									id: 'a3dbf468',
+									id: '834a887f',
 									key: {
 										dataType: DataTypes.String,
-										id: 'k8s_namespace_name--string--tag--false',
+										id: 'k8s_cluster_name--string--tag--false',
 										isColumn: false,
 										isJSON: false,
-										key: 'k8s_namespace_name',
+										key: 'k8s_cluster_name',
 										type: 'tag',
 									},
 									op: '=',
-									value: namespace.namespaceName,
+									value: cluster.meta.k8s_cluster_name,
 								},
 							],
 							op: 'AND',
@@ -474,7 +338,7 @@ export const getNamespaceQueryPayload = (
 						legend: 'usage (min)',
 						limit: null,
 						orderBy: [],
-						queryName: 'E',
+						queryName: 'B',
 						reduceTo: 'avg',
 						spaceAggregation: 'sum',
 						stepInterval: 60,
@@ -492,21 +356,21 @@ export const getNamespaceQueryPayload = (
 						aggregateOperator: 'max',
 						dataSource: DataSource.METRICS,
 						disabled: false,
-						expression: 'F',
+						expression: 'C',
 						filters: {
 							items: [
 								{
-									id: '4b2406c2',
+									id: '9b002160',
 									key: {
 										dataType: DataTypes.String,
-										id: 'k8s_namespace_name--string--tag--false',
+										id: 'k8s_cluster_name--string--tag--false',
 										isColumn: false,
 										isJSON: false,
-										key: 'k8s_namespace_name',
+										key: 'k8s_cluster_name',
 										type: 'tag',
 									},
 									op: '=',
-									value: namespace.namespaceName,
+									value: cluster.meta.k8s_cluster_name,
 								},
 							],
 							op: 'AND',
@@ -517,92 +381,50 @@ export const getNamespaceQueryPayload = (
 						legend: 'usage (max)',
 						limit: null,
 						orderBy: [],
-						queryName: 'F',
+						queryName: 'C',
 						reduceTo: 'avg',
 						spaceAggregation: 'sum',
 						stepInterval: 60,
 						timeAggregation: 'max',
 					},
-				],
-				queryFormulas: [],
-			},
-			clickhouse_sql: [
-				{
-					disabled: false,
-					legend: '',
-					name: 'A',
-					query: '',
-				},
-			],
-			id: v4(),
-			promql: [
-				{
-					disabled: false,
-					legend: '',
-					name: 'A',
-					query: '',
-				},
-			],
-			queryType: EQueryType.QUERY_BUILDER,
-		},
-		variables: {},
-		formatForWeb: false,
-		start,
-		end,
-	},
-	{
-		selectedTime: 'GLOBAL_TIME',
-		graphType: PANEL_TYPES.TIME_SERIES,
-		query: {
-			builder: {
-				queryData: [
 					{
 						aggregateAttribute: {
 							dataType: DataTypes.Float64,
-							id: 'k8s_pod_cpu_utilization--float64--Gauge--true',
+							id: 'k8s_node_allocatable_memory--float64--Gauge--true',
 							isColumn: true,
 							isJSON: false,
-							key: 'k8s_pod_cpu_utilization',
+							key: 'k8s_node_allocatable_memory',
 							type: 'Gauge',
 						},
 						aggregateOperator: 'avg',
 						dataSource: DataSource.METRICS,
 						disabled: false,
-						expression: 'A',
+						expression: 'D',
 						filters: {
 							items: [
 								{
-									id: 'c3a73f0a',
+									id: '02a9a67e',
 									key: {
 										dataType: DataTypes.String,
-										id: 'k8s_namespace_name--string--tag--false',
+										id: 'k8s_cluster_name--string--tag--false',
 										isColumn: false,
 										isJSON: false,
-										key: 'k8s_namespace_name',
+										key: 'k8s_cluster_name',
 										type: 'tag',
 									},
 									op: '=',
-									value: namespace.namespaceName,
+									value: cluster.meta.k8s_cluster_name,
 								},
 							],
 							op: 'AND',
 						},
 						functions: [],
-						groupBy: [
-							{
-								dataType: DataTypes.String,
-								id: 'k8s_pod_name--string--tag--false',
-								isColumn: false,
-								isJSON: false,
-								key: 'k8s_pod_name',
-								type: 'tag',
-							},
-						],
+						groupBy: [],
 						having: [],
-						legend: '{{k8s_pod_name}}',
-						limit: 20,
+						legend: 'allocatable',
+						limit: null,
 						orderBy: [],
-						queryName: 'A',
+						queryName: 'D',
 						reduceTo: 'avg',
 						spaceAggregation: 'sum',
 						stepInterval: 60,
@@ -644,30 +466,30 @@ export const getNamespaceQueryPayload = (
 					{
 						aggregateAttribute: {
 							dataType: DataTypes.Float64,
-							id: 'k8s_pod_memory_usage--float64--Gauge--true',
+							id: 'k8s_node_condition_ready--float64--Gauge--true',
 							isColumn: true,
 							isJSON: false,
-							key: 'k8s_pod_memory_usage',
+							key: 'k8s_node_condition_ready',
 							type: 'Gauge',
 						},
-						aggregateOperator: 'avg',
+						aggregateOperator: 'latest',
 						dataSource: DataSource.METRICS,
 						disabled: false,
 						expression: 'A',
 						filters: {
 							items: [
 								{
-									id: '5cad3379',
+									id: 'd7779183',
 									key: {
 										dataType: DataTypes.String,
-										id: 'k8s_namespace_name--string--tag--false',
+										id: 'k8s_cluster_name--string--tag--false',
 										isColumn: false,
 										isJSON: false,
-										key: 'k8s_namespace_name',
+										key: 'k8s_cluster_name',
 										type: 'tag',
 									},
 									op: '=',
-									value: namespace.namespaceName,
+									value: cluster.meta.k8s_cluster_name,
 								},
 							],
 							op: 'AND',
@@ -676,22 +498,28 @@ export const getNamespaceQueryPayload = (
 						groupBy: [
 							{
 								dataType: DataTypes.String,
-								id: 'k8s_pod_name--string--tag--false',
+								id: 'k8s_node_name--string--tag--false',
 								isColumn: false,
 								isJSON: false,
-								key: 'k8s_pod_name',
+								key: 'k8s_node_name',
 								type: 'tag',
 							},
 						],
-						having: [],
-						legend: '{{k8s_pod_name}}',
-						limit: 10,
+						having: [
+							{
+								columnName: 'MAX(k8s_node_condition_ready)',
+								op: '=',
+								value: 1,
+							},
+						],
+						legend: '{{k8s_node_name}}',
+						limit: null,
 						orderBy: [],
 						queryName: 'A',
 						reduceTo: 'avg',
-						spaceAggregation: 'sum',
+						spaceAggregation: 'max',
 						stepInterval: 60,
-						timeAggregation: 'avg',
+						timeAggregation: 'latest',
 					},
 				],
 				queryFormulas: [],
@@ -729,30 +557,30 @@ export const getNamespaceQueryPayload = (
 					{
 						aggregateAttribute: {
 							dataType: DataTypes.Float64,
-							id: 'k8s_pod_network_io--float64--Sum--true',
+							id: 'k8s_node_condition_ready--float64--Gauge--true',
 							isColumn: true,
 							isJSON: false,
-							key: 'k8s_pod_network_io',
-							type: 'Sum',
+							key: 'k8s_node_condition_ready',
+							type: 'Gauge',
 						},
-						aggregateOperator: 'rate',
+						aggregateOperator: 'latest',
 						dataSource: DataSource.METRICS,
 						disabled: false,
 						expression: 'A',
 						filters: {
 							items: [
 								{
-									id: '00f5c5e1',
+									id: 'd7779183',
 									key: {
 										dataType: DataTypes.String,
-										id: 'k8s_namespace_name--string--tag--false',
+										id: 'k8s_cluster_name--string--tag--false',
 										isColumn: false,
 										isJSON: false,
-										key: 'k8s_namespace_name',
+										key: 'k8s_cluster_name',
 										type: 'tag',
 									},
 									op: '=',
-									value: namespace.namespaceName,
+									value: cluster.meta.k8s_cluster_name,
 								},
 							],
 							op: 'AND',
@@ -761,30 +589,28 @@ export const getNamespaceQueryPayload = (
 						groupBy: [
 							{
 								dataType: DataTypes.String,
-								id: 'interface--string--tag--false',
+								id: 'k8s_node_name--string--tag--false',
 								isColumn: false,
 								isJSON: false,
-								key: 'interface',
-								type: 'tag',
-							},
-							{
-								dataType: DataTypes.String,
-								id: 'direction--string--tag--false',
-								isColumn: false,
-								isJSON: false,
-								key: 'direction',
+								key: 'k8s_node_name',
 								type: 'tag',
 							},
 						],
-						having: [],
-						legend: '{{direction}} :: {{interface}}',
+						having: [
+							{
+								columnName: 'MAX(k8s_node_condition_ready)',
+								op: '=',
+								value: 0,
+							},
+						],
+						legend: '{{k8s_node_name}}',
 						limit: null,
 						orderBy: [],
 						queryName: 'A',
 						reduceTo: 'avg',
-						spaceAggregation: 'sum',
+						spaceAggregation: 'max',
 						stepInterval: 60,
-						timeAggregation: 'rate',
+						timeAggregation: 'latest',
 					},
 				],
 				queryFormulas: [],
@@ -815,37 +641,37 @@ export const getNamespaceQueryPayload = (
 	},
 	{
 		selectedTime: 'GLOBAL_TIME',
-		graphType: PANEL_TYPES.TIME_SERIES,
+		graphType: PANEL_TYPES.TABLE,
 		query: {
 			builder: {
 				queryData: [
 					{
 						aggregateAttribute: {
-							dataType: DataTypes.String,
-							id: 'k8s_pod_network_errors--float64--Sum--true',
+							dataType: DataTypes.Float64,
+							id: 'k8s_deployment_available--float64--Gauge--true',
 							isColumn: true,
 							isJSON: false,
-							key: 'k8s_pod_network_errors',
-							type: 'Sum',
+							key: 'k8s_deployment_available',
+							type: 'Gauge',
 						},
-						aggregateOperator: 'increase',
+						aggregateOperator: 'latest',
 						dataSource: DataSource.METRICS,
 						disabled: false,
 						expression: 'A',
 						filters: {
 							items: [
 								{
-									id: '3aa8e064',
+									id: 'a7da59c7',
 									key: {
 										dataType: DataTypes.String,
-										id: 'k8s_namespace_name--string--tag--false',
+										id: 'k8s_cluster_name--string--tag--false',
 										isColumn: false,
 										isJSON: false,
-										key: 'k8s_namespace_name',
+										key: 'k8s_cluster_name',
 										type: 'tag',
 									},
 									op: '=',
-									value: namespace.namespaceName,
+									value: cluster.meta.k8s_cluster_name,
 								},
 							],
 							op: 'AND',
@@ -854,30 +680,90 @@ export const getNamespaceQueryPayload = (
 						groupBy: [
 							{
 								dataType: DataTypes.String,
-								id: 'interface--string--tag--false',
+								id: 'k8s_deployment_name--string--tag--false',
 								isColumn: false,
 								isJSON: false,
-								key: 'interface',
+								key: 'k8s_deployment_name',
 								type: 'tag',
 							},
 							{
 								dataType: DataTypes.String,
-								id: 'direction--string--tag--false',
+								id: 'k8s_namespace_name--string--tag--false',
 								isColumn: false,
 								isJSON: false,
-								key: 'direction',
+								key: 'k8s_namespace_name',
 								type: 'tag',
 							},
 						],
 						having: [],
-						legend: '{{direction}} :: {{interface}}',
+						legend: 'available',
 						limit: null,
 						orderBy: [],
 						queryName: 'A',
-						reduceTo: 'avg',
+						reduceTo: 'last',
 						spaceAggregation: 'sum',
 						stepInterval: 60,
-						timeAggregation: 'increase',
+						timeAggregation: 'latest',
+					},
+					{
+						aggregateAttribute: {
+							dataType: DataTypes.Float64,
+							id: 'k8s_deployment_desired--float64--Gauge--true',
+							isColumn: true,
+							isJSON: false,
+							key: 'k8s_deployment_desired',
+							type: 'Gauge',
+						},
+						aggregateOperator: 'avg',
+						dataSource: DataSource.METRICS,
+						disabled: false,
+						expression: 'B',
+						filters: {
+							items: [
+								{
+									id: '55110885',
+									key: {
+										dataType: DataTypes.String,
+										id: 'k8s_cluster_name--string--tag--false',
+										isColumn: false,
+										isJSON: false,
+										key: 'k8s_cluster_name',
+										type: 'tag',
+									},
+									op: '=',
+									value: cluster.meta.k8s_cluster_name,
+								},
+							],
+							op: 'AND',
+						},
+						functions: [],
+						groupBy: [
+							{
+								dataType: DataTypes.String,
+								id: 'k8s_deployment_name--string--tag--false',
+								isColumn: false,
+								isJSON: false,
+								key: 'k8s_deployment_name',
+								type: 'tag',
+							},
+							{
+								dataType: DataTypes.String,
+								id: 'k8s_namespace_name--string--tag--false',
+								isColumn: false,
+								isJSON: false,
+								key: 'k8s_namespace_name',
+								type: 'tag',
+							},
+						],
+						having: [],
+						legend: 'desired',
+						limit: null,
+						orderBy: [],
+						queryName: 'B',
+						reduceTo: 'last',
+						spaceAggregation: 'sum',
+						stepInterval: 60,
+						timeAggregation: 'avg',
 					},
 				],
 				queryFormulas: [],
@@ -902,7 +788,7 @@ export const getNamespaceQueryPayload = (
 			queryType: EQueryType.QUERY_BUILDER,
 		},
 		variables: {},
-		formatForWeb: false,
+		formatForWeb: true,
 		start,
 		end,
 	},
@@ -921,24 +807,24 @@ export const getNamespaceQueryPayload = (
 							key: 'k8s_statefulset_current_pods',
 							type: 'Gauge',
 						},
-						aggregateOperator: 'latest',
+						aggregateOperator: 'max',
 						dataSource: DataSource.METRICS,
 						disabled: false,
 						expression: 'A',
 						filters: {
 							items: [
 								{
-									id: '5f2a55c5',
+									id: '3c57b4d1',
 									key: {
 										dataType: DataTypes.String,
-										id: 'k8s_namespace_name--string--tag--false',
+										id: 'k8s_cluster_name--string--tag--false',
 										isColumn: false,
 										isJSON: false,
-										key: 'k8s_namespace_name',
+										key: 'k8s_cluster_name',
 										type: 'tag',
 									},
 									op: '=',
-									value: namespace.namespaceName,
+									value: cluster.meta.k8s_cluster_name,
 								},
 							],
 							op: 'AND',
@@ -953,6 +839,14 @@ export const getNamespaceQueryPayload = (
 								key: 'k8s_statefulset_name',
 								type: 'tag',
 							},
+							{
+								dataType: DataTypes.String,
+								id: 'k8s_namespace_name--string--tag--false',
+								isColumn: false,
+								isJSON: false,
+								key: 'k8s_namespace_name',
+								type: 'tag',
+							},
 						],
 						having: [],
 						legend: 'current',
@@ -960,9 +854,9 @@ export const getNamespaceQueryPayload = (
 						orderBy: [],
 						queryName: 'A',
 						reduceTo: 'last',
-						spaceAggregation: 'max',
+						spaceAggregation: 'sum',
 						stepInterval: 60,
-						timeAggregation: 'latest',
+						timeAggregation: 'max',
 					},
 					{
 						aggregateAttribute: {
@@ -973,24 +867,24 @@ export const getNamespaceQueryPayload = (
 							key: 'k8s_statefulset_desired_pods',
 							type: 'Gauge',
 						},
-						aggregateOperator: 'latest',
+						aggregateOperator: 'max',
 						dataSource: DataSource.METRICS,
 						disabled: false,
 						expression: 'B',
 						filters: {
 							items: [
 								{
-									id: '13bd7a1d',
+									id: '0f49fe64',
 									key: {
 										dataType: DataTypes.String,
-										id: 'k8s_namespace_name--string--tag--false',
+										id: 'k8s_cluster_name--string--tag--false',
 										isColumn: false,
 										isJSON: false,
-										key: 'k8s_namespace_name',
+										key: 'k8s_cluster_name',
 										type: 'tag',
 									},
 									op: '=',
-									value: namespace.namespaceName,
+									value: cluster.meta.k8s_cluster_name,
 								},
 							],
 							op: 'AND',
@@ -1005,6 +899,14 @@ export const getNamespaceQueryPayload = (
 								key: 'k8s_statefulset_name',
 								type: 'tag',
 							},
+							{
+								dataType: DataTypes.String,
+								id: 'k8s_namespace_name--string--tag--false',
+								isColumn: false,
+								isJSON: false,
+								key: 'k8s_namespace_name',
+								type: 'tag',
+							},
 						],
 						having: [],
 						legend: 'desired',
@@ -1012,9 +914,69 @@ export const getNamespaceQueryPayload = (
 						orderBy: [],
 						queryName: 'B',
 						reduceTo: 'last',
-						spaceAggregation: 'max',
+						spaceAggregation: 'sum',
 						stepInterval: 60,
-						timeAggregation: 'latest',
+						timeAggregation: 'max',
+					},
+					{
+						aggregateAttribute: {
+							dataType: DataTypes.Float64,
+							id: 'k8s_statefulset_ready_pods--float64--Gauge--true',
+							isColumn: true,
+							isJSON: false,
+							key: 'k8s_statefulset_ready_pods',
+							type: 'Gauge',
+						},
+						aggregateOperator: 'max',
+						dataSource: DataSource.METRICS,
+						disabled: false,
+						expression: 'C',
+						filters: {
+							items: [
+								{
+									id: '0bebf625',
+									key: {
+										dataType: DataTypes.String,
+										id: 'k8s_cluster_name--string--tag--false',
+										isColumn: false,
+										isJSON: false,
+										key: 'k8s_cluster_name',
+										type: 'tag',
+									},
+									op: '=',
+									value: cluster.meta.k8s_cluster_name,
+								},
+							],
+							op: 'AND',
+						},
+						functions: [],
+						groupBy: [
+							{
+								dataType: DataTypes.String,
+								id: 'k8s_statefulset_name--string--tag--false',
+								isColumn: false,
+								isJSON: false,
+								key: 'k8s_statefulset_name',
+								type: 'tag',
+							},
+							{
+								dataType: DataTypes.String,
+								id: 'k8s_namespace_name--string--tag--false',
+								isColumn: false,
+								isJSON: false,
+								key: 'k8s_namespace_name',
+								type: 'tag',
+							},
+						],
+						having: [],
+						legend: 'ready',
+						limit: null,
+						orderBy: [],
+						queryName: 'C',
+						reduceTo: 'last',
+						spaceAggregation: 'sum',
+						stepInterval: 60,
+						timeAggregation: 'max',
 					},
 					{
 						aggregateAttribute: {
@@ -1025,24 +987,24 @@ export const getNamespaceQueryPayload = (
 							key: 'k8s_statefulset_updated_pods',
 							type: 'Gauge',
 						},
-						aggregateOperator: 'latest',
+						aggregateOperator: 'max',
 						dataSource: DataSource.METRICS,
 						disabled: false,
-						expression: 'C',
+						expression: 'D',
 						filters: {
 							items: [
 								{
-									id: '9d287c73',
+									id: '1ddacbbe',
 									key: {
 										dataType: DataTypes.String,
-										id: 'k8s_namespace_name--string--tag--false',
+										id: 'k8s_cluster_name--string--tag--false',
 										isColumn: false,
 										isJSON: false,
-										key: 'k8s_namespace_name',
+										key: 'k8s_cluster_name',
 										type: 'tag',
 									},
 									op: '=',
-									value: namespace.namespaceName,
+									value: cluster.meta.k8s_cluster_name,
 								},
 							],
 							op: 'AND',
@@ -1057,406 +1019,24 @@ export const getNamespaceQueryPayload = (
 								key: 'k8s_statefulset_name',
 								type: 'tag',
 							},
+							{
+								dataType: DataTypes.String,
+								id: 'k8s_namespace_name--string--tag--false',
+								isColumn: false,
+								isJSON: false,
+								key: 'k8s_namespace_name',
+								type: 'tag',
+							},
 						],
 						having: [],
 						legend: 'updated',
 						limit: null,
 						orderBy: [],
-						queryName: 'C',
-						reduceTo: 'last',
-						spaceAggregation: 'max',
-						stepInterval: 60,
-						timeAggregation: 'latest',
-					},
-				],
-				queryFormulas: [],
-			},
-			clickhouse_sql: [
-				{
-					disabled: false,
-					legend: '',
-					name: 'A',
-					query: '',
-				},
-			],
-			id: v4(),
-			promql: [
-				{
-					disabled: false,
-					legend: '',
-					name: 'A',
-					query: '',
-				},
-			],
-			queryType: EQueryType.QUERY_BUILDER,
-		},
-		variables: {},
-		formatForWeb: true,
-		start,
-		end,
-	},
-	{
-		selectedTime: 'GLOBAL_TIME',
-		graphType: PANEL_TYPES.TABLE,
-		query: {
-			builder: {
-				queryData: [
-					{
-						aggregateAttribute: {
-							dataType: DataTypes.Float64,
-							id: 'k8s_replicaset_desired--float64--Gauge--true',
-							isColumn: true,
-							isJSON: false,
-							key: 'k8s_replicaset_desired',
-							type: 'Gauge',
-						},
-						aggregateOperator: 'latest',
-						dataSource: DataSource.METRICS,
-						disabled: false,
-						expression: 'A',
-						filters: {
-							items: [
-								{
-									id: '0c1e655c',
-									key: {
-										dataType: DataTypes.String,
-										id: 'k8s_namespace_name--string--tag--false',
-										isColumn: false,
-										isJSON: false,
-										key: 'k8s_namespace_name',
-										type: 'tag',
-									},
-									op: '=',
-									value: namespace.namespaceName,
-								},
-							],
-							op: 'AND',
-						},
-						functions: [],
-						groupBy: [
-							{
-								dataType: DataTypes.String,
-								id: 'k8s_replicaset_name--string--tag--false',
-								isColumn: false,
-								isJSON: false,
-								key: 'k8s_replicaset_name',
-								type: 'tag',
-							},
-						],
-						having: [
-							{
-								columnName: 'MAX(k8s_replicaset_desired)',
-								op: '>',
-								value: 0,
-							},
-						],
-						legend: 'desired',
-						limit: null,
-						orderBy: [],
-						queryName: 'A',
-						reduceTo: 'last',
-						spaceAggregation: 'max',
-						stepInterval: 60,
-						timeAggregation: 'latest',
-					},
-					{
-						aggregateAttribute: {
-							dataType: DataTypes.Float64,
-							id: 'k8s_replicaset_available--float64--Gauge--true',
-							isColumn: true,
-							isJSON: false,
-							key: 'k8s_replicaset_available',
-							type: 'Gauge',
-						},
-						aggregateOperator: 'latest',
-						dataSource: DataSource.METRICS,
-						disabled: false,
-						expression: 'B',
-						filters: {
-							items: [
-								{
-									id: 'b2296bdb',
-									key: {
-										dataType: DataTypes.String,
-										id: 'k8s_namespace_name--string--tag--false',
-										isColumn: false,
-										isJSON: false,
-										key: 'k8s_namespace_name',
-										type: 'tag',
-									},
-									op: '=',
-									value: namespace.namespaceName,
-								},
-							],
-							op: 'AND',
-						},
-						functions: [],
-						groupBy: [
-							{
-								dataType: DataTypes.String,
-								id: 'k8s_replicaset_name--string--tag--false',
-								isColumn: false,
-								isJSON: false,
-								key: 'k8s_replicaset_name',
-								type: 'tag',
-							},
-						],
-						having: [
-							{
-								columnName: 'MAX(k8s_replicaset_available)',
-								op: '>',
-								value: 0,
-							},
-						],
-						legend: 'available',
-						limit: null,
-						orderBy: [],
-						queryName: 'B',
-						reduceTo: 'last',
-						spaceAggregation: 'max',
-						stepInterval: 60,
-						timeAggregation: 'latest',
-					},
-				],
-				queryFormulas: [],
-			},
-			clickhouse_sql: [
-				{
-					disabled: false,
-					legend: '',
-					name: 'A',
-					query: '',
-				},
-			],
-			id: v4(),
-			promql: [
-				{
-					disabled: false,
-					legend: '',
-					name: 'A',
-					query: '',
-				},
-			],
-			queryType: EQueryType.QUERY_BUILDER,
-		},
-		variables: {},
-		formatForWeb: true,
-		start,
-		end,
-	},
-	{
-		selectedTime: 'GLOBAL_TIME',
-		graphType: PANEL_TYPES.TABLE,
-		query: {
-			builder: {
-				queryData: [
-					{
-						aggregateAttribute: {
-							dataType: DataTypes.Float64,
-							id: 'k8s_daemonset_desired_scheduled_namespaces--float64--Gauge--true',
-							isColumn: true,
-							isJSON: false,
-							key: 'k8s_daemonset_desired_scheduled_namespaces',
-							type: 'Gauge',
-						},
-						aggregateOperator: 'latest',
-						dataSource: DataSource.METRICS,
-						disabled: false,
-						expression: 'A',
-						filters: {
-							items: [
-								{
-									id: '2964eb92',
-									key: {
-										dataType: DataTypes.String,
-										id: 'k8s_namespace_name--string--tag--false',
-										isColumn: false,
-										isJSON: false,
-										key: 'k8s_namespace_name',
-										type: 'tag',
-									},
-									op: '=',
-									value: namespace.namespaceName,
-								},
-							],
-							op: 'AND',
-						},
-						functions: [],
-						groupBy: [
-							{
-								dataType: DataTypes.String,
-								id: 'k8s_daemonset_name--string--tag--false',
-								isColumn: false,
-								isJSON: false,
-								key: 'k8s_daemonset_name',
-								type: 'tag',
-							},
-						],
-						having: [],
-						legend: 'desired',
-						limit: null,
-						orderBy: [],
-						queryName: 'A',
-						reduceTo: 'last',
-						spaceAggregation: 'max',
-						stepInterval: 60,
-						timeAggregation: 'latest',
-					},
-					{
-						aggregateAttribute: {
-							dataType: DataTypes.Float64,
-							id: 'k8s_daemonset_current_scheduled_namespaces--float64--Gauge--true',
-							isColumn: true,
-							isJSON: false,
-							key: 'k8s_daemonset_current_scheduled_namespaces',
-							type: 'Gauge',
-						},
-						aggregateOperator: 'latest',
-						dataSource: DataSource.METRICS,
-						disabled: false,
-						expression: 'B',
-						filters: {
-							items: [
-								{
-									id: 'cd324eff',
-									key: {
-										dataType: DataTypes.String,
-										id: 'k8s_namespace_name--string--tag--false',
-										isColumn: false,
-										isJSON: false,
-										key: 'k8s_namespace_name',
-										type: 'tag',
-									},
-									op: '=',
-									value: namespace.namespaceName,
-								},
-							],
-							op: 'AND',
-						},
-						functions: [],
-						groupBy: [
-							{
-								dataType: DataTypes.String,
-								id: 'k8s_daemonset_name--string--tag--false',
-								isColumn: false,
-								isJSON: false,
-								key: 'k8s_daemonset_name',
-								type: 'tag',
-							},
-						],
-						having: [],
-						legend: 'current',
-						limit: null,
-						orderBy: [],
-						queryName: 'B',
-						reduceTo: 'last',
-						spaceAggregation: 'max',
-						stepInterval: 60,
-						timeAggregation: 'latest',
-					},
-					{
-						aggregateAttribute: {
-							dataType: DataTypes.Float64,
-							id: 'k8s_daemonset_ready_namespaces--float64--Gauge--true',
-							isColumn: true,
-							isJSON: false,
-							key: 'k8s_daemonset_ready_namespaces',
-							type: 'Gauge',
-						},
-						aggregateOperator: 'latest',
-						dataSource: DataSource.METRICS,
-						disabled: false,
-						expression: 'C',
-						filters: {
-							items: [
-								{
-									id: '0416fa6f',
-									key: {
-										dataType: DataTypes.String,
-										id: 'k8s_namespace_name--string--tag--false',
-										isColumn: false,
-										isJSON: false,
-										key: 'k8s_namespace_name',
-										type: 'tag',
-									},
-									op: '=',
-									value: namespace.namespaceName,
-								},
-							],
-							op: 'AND',
-						},
-						functions: [],
-						groupBy: [
-							{
-								dataType: DataTypes.String,
-								id: 'k8s_daemonset_name--string--tag--false',
-								isColumn: false,
-								isJSON: false,
-								key: 'k8s_daemonset_name',
-								type: 'tag',
-							},
-						],
-						having: [],
-						legend: 'ready',
-						limit: null,
-						orderBy: [],
-						queryName: 'C',
-						reduceTo: 'last',
-						spaceAggregation: 'max',
-						stepInterval: 60,
-						timeAggregation: 'latest',
-					},
-					{
-						aggregateAttribute: {
-							dataType: DataTypes.Float64,
-							id: 'k8s_daemonset_misscheduled_namespaces--float64--Gauge--true',
-							isColumn: true,
-							isJSON: false,
-							key: 'k8s_daemonset_misscheduled_namespaces',
-							type: 'Gauge',
-						},
-						aggregateOperator: 'latest',
-						dataSource: DataSource.METRICS,
-						disabled: false,
-						expression: 'D',
-						filters: {
-							items: [
-								{
-									id: 'c0a126d3',
-									key: {
-										dataType: DataTypes.String,
-										id: 'k8s_namespace_name--string--tag--false',
-										isColumn: false,
-										isJSON: false,
-										key: 'k8s_namespace_name',
-										type: 'tag',
-									},
-									op: '=',
-									value: namespace.namespaceName,
-								},
-							],
-							op: 'AND',
-						},
-						functions: [],
-						groupBy: [
-							{
-								dataType: DataTypes.String,
-								id: 'k8s_daemonset_name--string--tag--false',
-								isColumn: false,
-								isJSON: false,
-								key: 'k8s_daemonset_name',
-								type: 'tag',
-							},
-						],
-						having: [],
-						legend: 'misscheduled',
-						limit: null,
-						orderBy: [],
 						queryName: 'D',
 						reduceTo: 'last',
-						spaceAggregation: 'max',
+						spaceAggregation: 'sum',
 						stepInterval: 60,
-						timeAggregation: 'latest',
+						timeAggregation: 'max',
 					},
 				],
 				queryFormulas: [],
@@ -1494,30 +1074,30 @@ export const getNamespaceQueryPayload = (
 					{
 						aggregateAttribute: {
 							dataType: DataTypes.Float64,
-							id: 'k8s_deployment_desired--float64--Gauge--true',
+							id: 'k8s_daemonset_current_scheduled_nodes--float64--Gauge--true',
 							isColumn: true,
 							isJSON: false,
-							key: 'k8s_deployment_desired',
+							key: 'k8s_daemonset_current_scheduled_nodes',
 							type: 'Gauge',
 						},
-						aggregateOperator: 'latest',
+						aggregateOperator: 'avg',
 						dataSource: DataSource.METRICS,
 						disabled: false,
 						expression: 'A',
 						filters: {
 							items: [
 								{
-									id: '9bc659c1',
+									id: 'e0bea554',
 									key: {
 										dataType: DataTypes.String,
-										id: 'k8s_namespace_name--string--tag--false',
+										id: 'k8s_cluster_name--string--tag--false',
 										isColumn: false,
 										isJSON: false,
-										key: 'k8s_namespace_name',
+										key: 'k8s_cluster_name',
 										type: 'tag',
 									},
 									op: '=',
-									value: namespace.namespaceName,
+									value: cluster.meta.k8s_cluster_name,
 								},
 							],
 							op: 'AND',
@@ -1526,30 +1106,30 @@ export const getNamespaceQueryPayload = (
 						groupBy: [
 							{
 								dataType: DataTypes.String,
-								id: 'k8s_deployment_name--string--tag--false',
+								id: 'k8s_daemonset_name--string--tag--false',
 								isColumn: false,
 								isJSON: false,
-								key: 'k8s_deployment_name',
+								key: 'k8s_daemonset_name',
 								type: 'tag',
 							},
 						],
 						having: [],
-						legend: 'desired',
+						legend: 'current_nodes',
 						limit: null,
 						orderBy: [],
 						queryName: 'A',
 						reduceTo: 'last',
-						spaceAggregation: 'max',
+						spaceAggregation: 'avg',
 						stepInterval: 60,
-						timeAggregation: 'latest',
+						timeAggregation: 'avg',
 					},
 					{
 						aggregateAttribute: {
 							dataType: DataTypes.Float64,
-							id: 'k8s_deployment_available--float64--Gauge--true',
+							id: 'k8s_daemonset_desired_scheduled_nodes--float64--Gauge--true',
 							isColumn: true,
 							isJSON: false,
-							key: 'k8s_deployment_available',
+							key: 'k8s_daemonset_desired_scheduled_nodes',
 							type: 'Gauge',
 						},
 						aggregateOperator: 'avg',
@@ -1559,17 +1139,17 @@ export const getNamespaceQueryPayload = (
 						filters: {
 							items: [
 								{
-									id: 'e1696631',
+									id: '741052f7',
 									key: {
 										dataType: DataTypes.String,
-										id: 'k8s_namespace_name--string--tag--false',
+										id: 'k8s_cluster_name--string--tag--false',
 										isColumn: false,
 										isJSON: false,
-										key: 'k8s_namespace_name',
+										key: 'k8s_cluster_name',
 										type: 'tag',
 									},
 									op: '=',
-									value: namespace.namespaceName,
+									value: cluster.meta.k8s_cluster_name,
 								},
 							],
 							op: 'AND',
@@ -1578,32 +1158,350 @@ export const getNamespaceQueryPayload = (
 						groupBy: [
 							{
 								dataType: DataTypes.String,
-								id: 'k8s_deployment_name--string--tag--false',
+								id: 'k8s_daemonset_name--string--tag--false',
 								isColumn: false,
 								isJSON: false,
-								key: 'k8s_deployment_name',
+								key: 'k8s_daemonset_name',
 								type: 'tag',
 							},
 						],
 						having: [],
-						legend: 'available',
+						legend: 'desired_nodes',
 						limit: null,
 						orderBy: [],
 						queryName: 'B',
 						reduceTo: 'last',
-						spaceAggregation: 'max',
+						spaceAggregation: 'avg',
+						stepInterval: 60,
+						timeAggregation: 'avg',
+					},
+					{
+						aggregateAttribute: {
+							dataType: DataTypes.Float64,
+							id: 'k8s_daemonset_ready_nodes--float64--Gauge--true',
+							isColumn: true,
+							isJSON: false,
+							key: 'k8s_daemonset_ready_nodes',
+							type: 'Gauge',
+						},
+						aggregateOperator: 'avg',
+						dataSource: DataSource.METRICS,
+						disabled: false,
+						expression: 'C',
+						filters: {
+							items: [
+								{
+									id: 'f23759f2',
+									key: {
+										dataType: DataTypes.String,
+										id: 'k8s_cluster_name--string--tag--false',
+										isColumn: false,
+										isJSON: false,
+										key: 'k8s_cluster_name',
+										type: 'tag',
+									},
+									op: '=',
+									value: cluster.meta.k8s_cluster_name,
+								},
+							],
+							op: 'AND',
+						},
+						functions: [],
+						groupBy: [
+							{
+								dataType: DataTypes.String,
+								id: 'k8s_daemonset_name--string--tag--false',
+								isColumn: false,
+								isJSON: false,
+								key: 'k8s_daemonset_name',
+								type: 'tag',
+							},
+						],
+						having: [],
+						legend: 'ready_nodes',
+						limit: null,
+						orderBy: [],
+						queryName: 'C',
+						reduceTo: 'last',
+						spaceAggregation: 'avg',
 						stepInterval: 60,
 						timeAggregation: 'avg',
 					},
 				],
-				queryFormulas: [
+				queryFormulas: [],
+			},
+			clickhouse_sql: [
+				{
+					disabled: false,
+					legend: '',
+					name: 'A',
+					query: '',
+				},
+			],
+			id: v4(),
+			promql: [
+				{
+					disabled: false,
+					legend: '',
+					name: 'A',
+					query: '',
+				},
+			],
+			queryType: EQueryType.QUERY_BUILDER,
+		},
+		variables: {},
+		formatForWeb: true,
+		start,
+		end,
+	},
+	{
+		selectedTime: 'GLOBAL_TIME',
+		graphType: PANEL_TYPES.TABLE,
+		query: {
+			builder: {
+				queryData: [
 					{
+						aggregateAttribute: {
+							dataType: DataTypes.Float64,
+							id: 'k8s_job_active_pods--float64--Gauge--true',
+							isColumn: true,
+							isJSON: false,
+							key: 'k8s_job_active_pods',
+							type: 'Gauge',
+						},
+						aggregateOperator: 'max',
+						dataSource: DataSource.METRICS,
 						disabled: false,
-						expression: 'A/B',
-						legend: 'util %',
-						queryName: 'F1',
+						expression: 'A',
+						filters: {
+							items: [
+								{
+									id: 'fd485d85',
+									key: {
+										dataType: DataTypes.String,
+										id: 'k8s_cluster_name--string--tag--false',
+										isColumn: false,
+										isJSON: false,
+										key: 'k8s_cluster_name',
+										type: 'tag',
+									},
+									op: '=',
+									value: cluster.meta.k8s_cluster_name,
+								},
+							],
+							op: 'AND',
+						},
+						functions: [],
+						groupBy: [
+							{
+								dataType: DataTypes.String,
+								id: 'k8s_job_name--string--tag--false',
+								isColumn: false,
+								isJSON: false,
+								key: 'k8s_job_name',
+								type: 'tag',
+							},
+							{
+								dataType: DataTypes.String,
+								id: 'k8s_namespace_name--string--tag--false',
+								isColumn: false,
+								isJSON: false,
+								key: 'k8s_namespace_name',
+								type: 'tag',
+							},
+						],
+						having: [],
+						legend: 'running',
+						limit: null,
+						orderBy: [],
+						queryName: 'A',
+						reduceTo: 'last',
+						spaceAggregation: 'sum',
+						stepInterval: 60,
+						timeAggregation: 'max',
+					},
+					{
+						aggregateAttribute: {
+							dataType: DataTypes.Float64,
+							id: 'k8s_job_successful_pods--float64--Gauge--true',
+							isColumn: true,
+							isJSON: false,
+							key: 'k8s_job_successful_pods',
+							type: 'Gauge',
+						},
+						aggregateOperator: 'max',
+						dataSource: DataSource.METRICS,
+						disabled: false,
+						expression: 'B',
+						filters: {
+							items: [
+								{
+									id: 'fc1beb81',
+									key: {
+										dataType: DataTypes.String,
+										id: 'k8s_cluster_name--string--tag--false',
+										isColumn: false,
+										isJSON: false,
+										key: 'k8s_cluster_name',
+										type: 'tag',
+									},
+									op: '=',
+									value: cluster.meta.k8s_cluster_name,
+								},
+							],
+							op: 'AND',
+						},
+						functions: [],
+						groupBy: [
+							{
+								dataType: DataTypes.String,
+								id: 'k8s_job_name--string--tag--false',
+								isColumn: false,
+								isJSON: false,
+								key: 'k8s_job_name',
+								type: 'tag',
+							},
+							{
+								dataType: DataTypes.String,
+								id: 'k8s_namespace_name--string--tag--false',
+								isColumn: false,
+								isJSON: false,
+								key: 'k8s_namespace_name',
+								type: 'tag',
+							},
+						],
+						having: [],
+						legend: 'successful',
+						limit: null,
+						orderBy: [],
+						queryName: 'B',
+						reduceTo: 'last',
+						spaceAggregation: 'sum',
+						stepInterval: 60,
+						timeAggregation: 'max',
+					},
+					{
+						aggregateAttribute: {
+							dataType: DataTypes.Float64,
+							id: 'k8s_job_failed_pods--float64--Gauge--true',
+							isColumn: true,
+							isJSON: false,
+							key: 'k8s_job_failed_pods',
+							type: 'Gauge',
+						},
+						aggregateOperator: 'max',
+						dataSource: DataSource.METRICS,
+						disabled: false,
+						expression: 'C',
+						filters: {
+							items: [
+								{
+									id: '97773348',
+									key: {
+										dataType: DataTypes.String,
+										id: 'k8s_cluster_name--string--tag--false',
+										isColumn: false,
+										isJSON: false,
+										key: 'k8s_cluster_name',
+										type: 'tag',
+									},
+									op: '=',
+									value: cluster.meta.k8s_cluster_name,
+								},
+							],
+							op: 'AND',
+						},
+						functions: [],
+						groupBy: [
+							{
+								dataType: DataTypes.String,
+								id: 'k8s_job_name--string--tag--false',
+								isColumn: false,
+								isJSON: false,
+								key: 'k8s_job_name',
+								type: 'tag',
+							},
+							{
+								dataType: DataTypes.String,
+								id: 'k8s_namespace_name--string--tag--false',
+								isColumn: false,
+								isJSON: false,
+								key: 'k8s_namespace_name',
+								type: 'tag',
+							},
+						],
+						having: [],
+						legend: 'failed',
+						limit: null,
+						orderBy: [],
+						queryName: 'C',
+						reduceTo: 'last',
+						spaceAggregation: 'sum',
+						stepInterval: 60,
+						timeAggregation: 'max',
+					},
+					{
+						aggregateAttribute: {
+							dataType: DataTypes.Float64,
+							id: 'k8s_job_desired_successful_pods--float64--Gauge--true',
+							isColumn: true,
+							isJSON: false,
+							key: 'k8s_job_desired_successful_pods',
+							type: 'Gauge',
+						},
+						aggregateOperator: 'max',
+						dataSource: DataSource.METRICS,
+						disabled: false,
+						expression: 'D',
+						filters: {
+							items: [
+								{
+									id: '5911618b',
+									key: {
+										dataType: DataTypes.String,
+										id: 'k8s_cluster_name--string--tag--false',
+										isColumn: false,
+										isJSON: false,
+										key: 'k8s_cluster_name',
+										type: 'tag',
+									},
+									op: '=',
+									value: cluster.meta.k8s_cluster_name,
+								},
+							],
+							op: 'AND',
+						},
+						functions: [],
+						groupBy: [
+							{
+								dataType: DataTypes.String,
+								id: 'k8s_job_name--string--tag--false',
+								isColumn: false,
+								isJSON: false,
+								key: 'k8s_job_name',
+								type: 'tag',
+							},
+							{
+								dataType: DataTypes.String,
+								id: 'k8s_namespace_name--string--tag--false',
+								isColumn: false,
+								isJSON: false,
+								key: 'k8s_namespace_name',
+								type: 'tag',
+							},
+						],
+						having: [],
+						legend: 'desired successful',
+						limit: null,
+						orderBy: [],
+						queryName: 'D',
+						reduceTo: 'last',
+						spaceAggregation: 'sum',
+						stepInterval: 60,
+						timeAggregation: 'max',
 					},
 				],
+				queryFormulas: [],
 			},
 			clickhouse_sql: [
 				{
