--- frontend/src/container/InfraMonitoringK8s/Deployments/DeploymentDetails/Metrics/index.ts
@@ -1,3 +1,0 @@
-import DeploymentMetrics from './DeploymentMetrics';
-
-export default DeploymentMetrics;
