--- frontend/src/container/InfraMonitoringK8s/Namespaces/NamespaceDetails/Metrics/index.ts
@@ -1,3 +1,0 @@
-import NamespaceMetrics from './NamespaceMetrics';
-
-export default NamespaceMetrics;
