+++ frontend/src/container/InfraMonitoringK8s/Namespaces/NamespaceDetails/constants.ts
@@ -1,5 +1,5 @@
 /* eslint-disable sonarjs/no-duplicate-string */
-import { K8sClustersData } from 'api/infraMonitoring/getK8sClustersList';
+import { K8sNamespacesData } from 'api/infraMonitoring/getK8sNamespacesList';
 import { PANEL_TYPES } from 'constants/queryBuilder';
 import { GetQueryResultsProps } from 'lib/dashboard/getQueryResults';
 import { DataTypes } from 'types/api/queryBuilder/queryAutocompleteResponse';
@@ -7,43 +7,51 @@ import { EQueryType } from 'types/common/dashboard';
 import { DataSource } from 'types/common/queryBuilder';
 import { v4 } from 'uuid';
 
-export const clusterWidgetInfo = [
+export const namespaceWidgetInfo = [
 	{
-		title: 'CPU Usage, allocatable',
+		title: 'CPU Usage (cores)',
 		yAxisUnit: '',
 	},
 	{
-		title: 'Memory Usage, allocatable',
+		title: 'Memory Usage (bytes)',
 		yAxisUnit: 'bytes',
 	},
 	{
-		title: 'Ready Nodes',
+		title: 'Pods CPU (top 10)',
 		yAxisUnit: '',
 	},
 	{
-		title: 'NotReady Nodes',
+		title: 'Pods Memory (top 10)',
+		yAxisUnit: 'bytes',
+	},
+	{
+		title: 'Network rate',
+		yAxisUnit: 'binBps',
+	},
+	{
+		title: 'Network errors',
 		yAxisUnit: '',
 	},
 	{
-		title: 'Deployments available and desired',
+		title: 'StatefulSets',
 		yAxisUnit: '',
 	},
 	{
-		title: 'Statefulset pods',
+		title: 'ReplicaSets',
 		yAxisUnit: '',
 	},
 	{
-		title: 'Daemonset nodes',
+		title: 'DaemonSets',
 		yAxisUnit: '',
 	},
 	{
-		title: 'Jobs',
+		title: 'Deployments',
 		yAxisUnit: '',
 	},
 ];
 
-export const getClusterQueryPayload = (
-	cluster: K8sClustersData,
+export const getNamespaceMetricsQueryPayload = (
+	namespace: K8sNamespacesData,
 	start: number,
 	end: number,
 ): GetQueryResultsProps[] => [
@@ -69,17 +77,17 @@ export const getClusterQueryPayload = (
 						filters: {
 							items: [
 								{
-									id: '0224c582',
+									id: '47b3adae',
 									key: {
 										dataType: DataTypes.String,
-										id: 'k8s_cluster_name--string--tag--false',
+										id: 'k8s_namespace_name--string--tag--false',
 										isColumn: false,
 										isJSON: false,
-										key: 'k8s_cluster_name',
+										key: 'k8s_namespace_name',
 										type: 'tag',
 									},
 									op: '=',
-									value: cluster.meta.k8s_cluster_name,
+									value: namespace.namespaceName,
 								},
 							],
 							op: 'AND',
@@ -99,6 +107,49 @@ export const getClusterQueryPayload = (
 					{
 						aggregateAttribute: {
 							dataType: DataTypes.Float64,
+							id: 'k8s_container_cpu_request--float64--Gauge--true',
+							isColumn: true,
+							isJSON: false,
+							key: 'k8s_container_cpu_request',
+							type: 'Gauge',
+						},
+						aggregateOperator: 'latest',
+						dataSource: DataSource.METRICS,
+						disabled: false,
+						expression: 'B',
+						filters: {
+							items: [
+								{
+									id: '93d2be5e',
+									key: {
+										dataType: DataTypes.String,
+										id: 'k8s_namespace_name--string--tag--false',
+										isColumn: false,
+										isJSON: false,
+										key: 'k8s_namespace_name',
+										type: 'tag',
+									},
+									op: '=',
+									value: namespace.namespaceName,
+								},
+							],
+							op: 'AND',
+						},
+						functions: [],
+						groupBy: [],
+						having: [],
+						legend: 'requests',
+						limit: null,
+						orderBy: [],
+						queryName: 'B',
+						reduceTo: 'avg',
+						spaceAggregation: 'sum',
+						stepInterval: 60,
+						timeAggregation: 'latest',
+					},
+					{
+						aggregateAttribute: {
+							dataType: DataTypes.Float64,
 							id: 'k8s_pod_cpu_utilization--float64--Gauge--true',
 							isColumn: true,
 							isJSON: false,
@@ -108,21 +159,21 @@ export const getClusterQueryPayload = (
 						aggregateOperator: 'min',
 						dataSource: DataSource.METRICS,
 						disabled: false,
-						expression: 'B',
+						expression: 'C',
 						filters: {
 							items: [
 								{
-									id: 'a0e11f0c',
+									id: '795eb679',
 									key: {
 										dataType: DataTypes.String,
-										id: 'k8s_cluster_name--string--tag--false',
+										id: 'k8s_namespace_name--string--tag--false',
 										isColumn: false,
 										isJSON: false,
-										key: 'k8s_cluster_name',
+										key: 'k8s_namespace_name',
 										type: 'tag',
 									},
 									op: '=',
-									value: cluster.meta.k8s_cluster_name,
+									value: namespace.namespaceName,
 								},
 							],
 							op: 'AND',
@@ -133,7 +184,7 @@ export const getClusterQueryPayload = (
 						legend: 'usage (min)',
 						limit: null,
 						orderBy: [],
-						queryName: 'B',
+						queryName: 'C',
 						reduceTo: 'avg',
 						spaceAggregation: 'sum',
 						stepInterval: 60,
@@ -151,21 +202,21 @@ export const getClusterQueryPayload = (
 						aggregateOperator: 'max',
 						dataSource: DataSource.METRICS,
 						disabled: false,
-						expression: 'C',
+						expression: 'D',
 						filters: {
 							items: [
 								{
-									id: 'c775629c',
+									id: '6792adbe',
 									key: {
 										dataType: DataTypes.String,
-										id: 'k8s_cluster_name--string--tag--false',
+										id: 'k8s_namespace_name--string--tag--false',
 										isColumn: false,
 										isJSON: false,
-										key: 'k8s_cluster_name',
+										key: 'k8s_namespace_name',
 										type: 'tag',
 									},
 									op: '=',
-									value: cluster.meta.k8s_cluster_name,
+									value: namespace.namespaceName,
 								},
 							],
 							op: 'AND',
@@ -176,54 +227,11 @@ export const getClusterQueryPayload = (
 						legend: 'usage (max)',
 						limit: null,
 						orderBy: [],
-						queryName: 'C',
-						reduceTo: 'avg',
-						spaceAggregation: 'sum',
-						stepInterval: 60,
-						timeAggregation: 'max',
-					},
-					{
-						aggregateAttribute: {
-							dataType: DataTypes.Float64,
-							id: 'k8s_node_allocatable_cpu--float64--Gauge--true',
-							isColumn: true,
-							isJSON: false,
-							key: 'k8s_node_allocatable_cpu',
-							type: 'Gauge',
-						},
-						aggregateOperator: 'avg',
-						dataSource: DataSource.METRICS,
-						disabled: false,
-						expression: 'D',
-						filters: {
-							items: [
-								{
-									id: '31f6c43a',
-									key: {
-										dataType: DataTypes.String,
-										id: 'k8s_cluster_name--string--tag--false',
-										isColumn: false,
-										isJSON: false,
-										key: 'k8s_cluster_name',
-										type: 'tag',
-									},
-									op: '=',
-									value: cluster.meta.k8s_cluster_name,
-								},
-							],
-							op: 'AND',
-						},
-						functions: [],
-						groupBy: [],
-						having: [],
-						legend: 'allocatable',
-						limit: null,
-						orderBy: [],
 						queryName: 'D',
 						reduceTo: 'avg',
 						spaceAggregation: 'sum',
 						stepInterval: 60,
-						timeAggregation: 'avg',
+						timeAggregation: 'max',
 					},
 				],
 				queryFormulas: [],
@@ -274,17 +282,17 @@ export const getClusterQueryPayload = (
 						filters: {
 							items: [
 								{
-									id: '61a3c55e',
+									id: '10011298',
 									key: {
 										dataType: DataTypes.String,
-										id: 'k8s_cluster_name--string--tag--false',
+										id: 'k8s_namespace_name--string--tag--false',
 										isColumn: false,
 										isJSON: false,
-										key: 'k8s_cluster_name',
+										key: 'k8s_namespace_name',
 										type: 'tag',
 									},
 									op: '=',
-									value: cluster.meta.k8s_cluster_name,
+									value: namespace.namespaceName,
 								},
 							],
 							op: 'AND',
@@ -304,6 +312,135 @@ export const getClusterQueryPayload = (
 					{
 						aggregateAttribute: {
 							dataType: DataTypes.Float64,
+							id: 'k8s_container_memory_request--float64--Gauge--true',
+							isColumn: true,
+							isJSON: false,
+							key: 'k8s_container_memory_request',
+							type: 'Gauge',
+						},
+						aggregateOperator: 'latest',
+						dataSource: DataSource.METRICS,
+						disabled: false,
+						expression: 'B',
+						filters: {
+							items: [
+								{
+									id: 'ea53b656',
+									key: {
+										dataType: DataTypes.String,
+										id: 'k8s_namespace_name--string--tag--false',
+										isColumn: false,
+										isJSON: false,
+										key: 'k8s_namespace_name',
+										type: 'tag',
+									},
+									op: '=',
+									value: namespace.namespaceName,
+								},
+							],
+							op: 'AND',
+						},
+						functions: [],
+						groupBy: [],
+						having: [],
+						legend: 'request',
+						limit: null,
+						orderBy: [],
+						queryName: 'B',
+						reduceTo: 'avg',
+						spaceAggregation: 'sum',
+						stepInterval: 60,
+						timeAggregation: 'latest',
+					},
+					{
+						aggregateAttribute: {
+							dataType: DataTypes.Float64,
+							id: 'k8s_pod_memory_working_set--float64--Gauge--true',
+							isColumn: true,
+							isJSON: false,
+							key: 'k8s_pod_memory_working_set',
+							type: 'Gauge',
+						},
+						aggregateOperator: 'avg',
+						dataSource: DataSource.METRICS,
+						disabled: false,
+						expression: 'C',
+						filters: {
+							items: [
+								{
+									id: '674ace83',
+									key: {
+										dataType: DataTypes.String,
+										id: 'k8s_namespace_name--string--tag--false',
+										isColumn: false,
+										isJSON: false,
+										key: 'k8s_namespace_name',
+										type: 'tag',
+									},
+									op: '=',
+									value: namespace.namespaceName,
+								},
+							],
+							op: 'AND',
+						},
+						functions: [],
+						groupBy: [],
+						having: [],
+						legend: 'working set',
+						limit: null,
+						orderBy: [],
+						queryName: 'C',
+						reduceTo: 'avg',
+						spaceAggregation: 'sum',
+						stepInterval: 60,
+						timeAggregation: 'avg',
+					},
+					{
+						aggregateAttribute: {
+							dataType: DataTypes.Float64,
+							id: 'k8s_pod_memory_rss--float64--Gauge--true',
+							isColumn: true,
+							isJSON: false,
+							key: 'k8s_pod_memory_rss',
+							type: 'Gauge',
+						},
+						aggregateOperator: 'avg',
+						dataSource: DataSource.METRICS,
+						disabled: false,
+						expression: 'D',
+						filters: {
+							items: [
+								{
+									id: '187dbdb3',
+									key: {
+										dataType: DataTypes.String,
+										id: 'k8s_namespace_name--string--tag--false',
+										isColumn: false,
+										isJSON: false,
+										key: 'k8s_namespace_name',
+										type: 'tag',
+									},
+									op: '=',
+									value: namespace.namespaceName,
+								},
+							],
+							op: 'AND',
+						},
+						functions: [],
+						groupBy: [],
+						having: [],
+						legend: 'rss',
+						limit: null,
+						orderBy: [],
+						queryName: 'D',
+						reduceTo: 'avg',
+						spaceAggregation: 'sum',
+						stepInterval: 60,
+						timeAggregation: 'avg',
+					},
+					{
+						aggregateAttribute: {
+							dataType: DataTypes.Float64,
 							id: 'k8s_pod_memory_usage--float64--Gauge--true',
 							isColumn: true,
 							isJSON: false,
@@ -313,21 +450,21 @@ export const getClusterQueryPayload = (
 						aggregateOperator: 'min',
 						dataSource: DataSource.METRICS,
 						disabled: false,
-						expression: 'B',
+						expression: 'E',
 						filters: {
 							items: [
 								{
-									id: '834a887f',
+									id: 'a3dbf468',
 									key: {
 										dataType: DataTypes.String,
-										id: 'k8s_cluster_name--string--tag--false',
+										id: 'k8s_namespace_name--string--tag--false',
 										isColumn: false,
 										isJSON: false,
-										key: 'k8s_cluster_name',
+										key: 'k8s_namespace_name',
 										type: 'tag',
 									},
 									op: '=',
-									value: cluster.meta.k8s_cluster_name,
+									value: namespace.namespaceName,
 								},
 							],
 							op: 'AND',
@@ -338,7 +475,7 @@ export const getClusterQueryPayload = (
 						legend: 'usage (min)',
 						limit: null,
 						orderBy: [],
-						queryName: 'B',
+						queryName: 'E',
 						reduceTo: 'avg',
 						spaceAggregation: 'sum',
 						stepInterval: 60,
@@ -356,21 +493,21 @@ export const getClusterQueryPayload = (
 						aggregateOperator: 'max',
 						dataSource: DataSource.METRICS,
 						disabled: false,
-						expression: 'C',
+						expression: 'F',
 						filters: {
 							items: [
 								{
-									id: '9b002160',
+									id: '4b2406c2',
 									key: {
 										dataType: DataTypes.String,
-										id: 'k8s_cluster_name--string--tag--false',
+										id: 'k8s_namespace_name--string--tag--false',
 										isColumn: false,
 										isJSON: false,
-										key: 'k8s_cluster_name',
+										key: 'k8s_namespace_name',
 										type: 'tag',
 									},
 									op: '=',
-									value: cluster.meta.k8s_cluster_name,
+									value: namespace.namespaceName,
 								},
 							],
 							op: 'AND',
@@ -381,50 +518,92 @@ export const getClusterQueryPayload = (
 						legend: 'usage (max)',
 						limit: null,
 						orderBy: [],
-						queryName: 'C',
+						queryName: 'F',
 						reduceTo: 'avg',
 						spaceAggregation: 'sum',
 						stepInterval: 60,
 						timeAggregation: 'max',
 					},
+				],
+				queryFormulas: [],
+			},
+			clickhouse_sql: [
+				{
+					disabled: false,
+					legend: '',
+					name: 'A',
+					query: '',
+				},
+			],
+			id: v4(),
+			promql: [
+				{
+					disabled: false,
+					legend: '',
+					name: 'A',
+					query: '',
+				},
+			],
+			queryType: EQueryType.QUERY_BUILDER,
+		},
+		variables: {},
+		formatForWeb: false,
+		start,
+		end,
+	},
+	{
+		selectedTime: 'GLOBAL_TIME',
+		graphType: PANEL_TYPES.TIME_SERIES,
+		query: {
+			builder: {
+				queryData: [
 					{
 						aggregateAttribute: {
 							dataType: DataTypes.Float64,
-							id: 'k8s_node_allocatable_memory--float64--Gauge--true',
+							id: 'k8s_pod_cpu_utilization--float64--Gauge--true',
 							isColumn: true,
 							isJSON: false,
-							key: 'k8s_node_allocatable_memory',
+							key: 'k8s_pod_cpu_utilization',
 							type: 'Gauge',
 						},
 						aggregateOperator: 'avg',
 						dataSource: DataSource.METRICS,
 						disabled: false,
-						expression: 'D',
+						expression: 'A',
 						filters: {
 							items: [
 								{
-									id: '02a9a67e',
+									id: 'c3a73f0a',
 									key: {
 										dataType: DataTypes.String,
-										id: 'k8s_cluster_name--string--tag--false',
+										id: 'k8s_namespace_name--string--tag--false',
 										isColumn: false,
 										isJSON: false,
-										key: 'k8s_cluster_name',
+										key: 'k8s_namespace_name',
 										type: 'tag',
 									},
 									op: '=',
-									value: cluster.meta.k8s_cluster_name,
+									value: namespace.namespaceName,
 								},
 							],
 							op: 'AND',
 						},
 						functions: [],
-						groupBy: [],
+						groupBy: [
+							{
+								dataType: DataTypes.String,
+								id: 'k8s_pod_name--string--tag--false',
+								isColumn: false,
+								isJSON: false,
+								key: 'k8s_pod_name',
+								type: 'tag',
+							},
+						],
 						having: [],
-						legend: 'allocatable',
-						limit: null,
+						legend: '{{k8s_pod_name}}',
+						limit: 20,
 						orderBy: [],
-						queryName: 'D',
+						queryName: 'A',
 						reduceTo: 'avg',
 						spaceAggregation: 'sum',
 						stepInterval: 60,
@@ -466,30 +645,30 @@ export const getClusterQueryPayload = (
 					{
 						aggregateAttribute: {
 							dataType: DataTypes.Float64,
-							id: 'k8s_node_condition_ready--float64--Gauge--true',
+							id: 'k8s_pod_memory_usage--float64--Gauge--true',
 							isColumn: true,
 							isJSON: false,
-							key: 'k8s_node_condition_ready',
+							key: 'k8s_pod_memory_usage',
 							type: 'Gauge',
 						},
-						aggregateOperator: 'latest',
+						aggregateOperator: 'avg',
 						dataSource: DataSource.METRICS,
 						disabled: false,
 						expression: 'A',
 						filters: {
 							items: [
 								{
-									id: 'd7779183',
+									id: '5cad3379',
 									key: {
 										dataType: DataTypes.String,
-										id: 'k8s_cluster_name--string--tag--false',
+										id: 'k8s_namespace_name--string--tag--false',
 										isColumn: false,
 										isJSON: false,
-										key: 'k8s_cluster_name',
+										key: 'k8s_namespace_name',
 										type: 'tag',
 									},
 									op: '=',
-									value: cluster.meta.k8s_cluster_name,
+									value: namespace.namespaceName,
 								},
 							],
 							op: 'AND',
@@ -498,28 +677,22 @@ export const getClusterQueryPayload = (
 						groupBy: [
 							{
 								dataType: DataTypes.String,
-								id: 'k8s_node_name--string--tag--false',
+								id: 'k8s_pod_name--string--tag--false',
 								isColumn: false,
 								isJSON: false,
-								key: 'k8s_node_name',
+								key: 'k8s_pod_name',
 								type: 'tag',
 							},
 						],
-						having: [
-							{
-								columnName: 'MAX(k8s_node_condition_ready)',
-								op: '=',
-								value: 1,
-							},
-						],
-						legend: '{{k8s_node_name}}',
-						limit: null,
+						having: [],
+						legend: '{{k8s_pod_name}}',
+						limit: 10,
 						orderBy: [],
 						queryName: 'A',
 						reduceTo: 'avg',
-						spaceAggregation: 'max',
+						spaceAggregation: 'sum',
 						stepInterval: 60,
-						timeAggregation: 'latest',
+						timeAggregation: 'avg',
 					},
 				],
 				queryFormulas: [],
@@ -557,30 +730,30 @@ export const getClusterQueryPayload = (
 					{
 						aggregateAttribute: {
 							dataType: DataTypes.Float64,
-							id: 'k8s_node_condition_ready--float64--Gauge--true',
+							id: 'k8s_pod_network_io--float64--Sum--true',
 							isColumn: true,
 							isJSON: false,
-							key: 'k8s_node_condition_ready',
-							type: 'Gauge',
+							key: 'k8s_pod_network_io',
+							type: 'Sum',
 						},
-						aggregateOperator: 'latest',
+						aggregateOperator: 'rate',
 						dataSource: DataSource.METRICS,
 						disabled: false,
 						expression: 'A',
 						filters: {
 							items: [
 								{
-									id: 'd7779183',
+									id: '00f5c5e1',
 									key: {
 										dataType: DataTypes.String,
-										id: 'k8s_cluster_name--string--tag--false',
+										id: 'k8s_namespace_name--string--tag--false',
 										isColumn: false,
 										isJSON: false,
-										key: 'k8s_cluster_name',
+										key: 'k8s_namespace_name',
 										type: 'tag',
 									},
 									op: '=',
-									value: cluster.meta.k8s_cluster_name,
+									value: namespace.namespaceName,
 								},
 							],
 							op: 'AND',
@@ -589,28 +762,30 @@ export const getClusterQueryPayload = (
 						groupBy: [
 							{
 								dataType: DataTypes.String,
-								id: 'k8s_node_name--string--tag--false',
+								id: 'interface--string--tag--false',
 								isColumn: false,
 								isJSON: false,
-								key: 'k8s_node_name',
+								key: 'interface',
+								type: 'tag',
+							},
+							{
+								dataType: DataTypes.String,
+								id: 'direction--string--tag--false',
+								isColumn: false,
+								isJSON: false,
+								key: 'direction',
 								type: 'tag',
 							},
 						],
-						having: [
-							{
-								columnName: 'MAX(k8s_node_condition_ready)',
-								op: '=',
-								value: 0,
-							},
-						],
-						legend: '{{k8s_node_name}}',
+						having: [],
+						legend: '{{direction}} :: {{interface}}',
 						limit: null,
 						orderBy: [],
 						queryName: 'A',
 						reduceTo: 'avg',
-						spaceAggregation: 'max',
+						spaceAggregation: 'sum',
 						stepInterval: 60,
-						timeAggregation: 'latest',
+						timeAggregation: 'rate',
 					},
 				],
 				queryFormulas: [],
@@ -641,37 +816,37 @@ export const getClusterQueryPayload = (
 	},
 	{
 		selectedTime: 'GLOBAL_TIME',
-		graphType: PANEL_TYPES.TABLE,
+		graphType: PANEL_TYPES.TIME_SERIES,
 		query: {
 			builder: {
 				queryData: [
 					{
 						aggregateAttribute: {
-							dataType: DataTypes.Float64,
-							id: 'k8s_deployment_available--float64--Gauge--true',
+							dataType: DataTypes.String,
+							id: 'k8s_pod_network_errors--float64--Sum--true',
 							isColumn: true,
 							isJSON: false,
-							key: 'k8s_deployment_available',
-							type: 'Gauge',
+							key: 'k8s_pod_network_errors',
+							type: 'Sum',
 						},
-						aggregateOperator: 'latest',
+						aggregateOperator: 'increase',
 						dataSource: DataSource.METRICS,
 						disabled: false,
 						expression: 'A',
 						filters: {
 							items: [
 								{
-									id: 'a7da59c7',
+									id: '3aa8e064',
 									key: {
 										dataType: DataTypes.String,
-										id: 'k8s_cluster_name--string--tag--false',
+										id: 'k8s_namespace_name--string--tag--false',
 										isColumn: false,
 										isJSON: false,
-										key: 'k8s_cluster_name',
+										key: 'k8s_namespace_name',
 										type: 'tag',
 									},
 									op: '=',
-									value: cluster.meta.k8s_cluster_name,
+									value: namespace.namespaceName,
 								},
 							],
 							op: 'AND',
@@ -680,90 +855,30 @@ export const getClusterQueryPayload = (
 						groupBy: [
 							{
 								dataType: DataTypes.String,
-								id: 'k8s_deployment_name--string--tag--false',
+								id: 'interface--string--tag--false',
 								isColumn: false,
 								isJSON: false,
-								key: 'k8s_deployment_name',
+								key: 'interface',
 								type: 'tag',
 							},
 							{
 								dataType: DataTypes.String,
-								id: 'k8s_namespace_name--string--tag--false',
+								id: 'direction--string--tag--false',
 								isColumn: false,
 								isJSON: false,
-								key: 'k8s_namespace_name',
+								key: 'direction',
 								type: 'tag',
 							},
 						],
 						having: [],
-						legend: 'available',
+						legend: '{{direction}} :: {{interface}}',
 						limit: null,
 						orderBy: [],
 						queryName: 'A',
-						reduceTo: 'last',
+						reduceTo: 'avg',
 						spaceAggregation: 'sum',
 						stepInterval: 60,
-						timeAggregation: 'latest',
-					},
-					{
-						aggregateAttribute: {
-							dataType: DataTypes.Float64,
-							id: 'k8s_deployment_desired--float64--Gauge--true',
-							isColumn: true,
-							isJSON: false,
-							key: 'k8s_deployment_desired',
-							type: 'Gauge',
-						},
-						aggregateOperator: 'avg',
-						dataSource: DataSource.METRICS,
-						disabled: false,
-						expression: 'B',
-						filters: {
-							items: [
-								{
-									id: '55110885',
-									key: {
-										dataType: DataTypes.String,
-										id: 'k8s_cluster_name--string--tag--false',
-										isColumn: false,
-										isJSON: false,
-										key: 'k8s_cluster_name',
-										type: 'tag',
-									},
-									op: '=',
-									value: cluster.meta.k8s_cluster_name,
-								},
-							],
-							op: 'AND',
-						},
-						functions: [],
-						groupBy: [
-							{
-								dataType: DataTypes.String,
-								id: 'k8s_deployment_name--string--tag--false',
-								isColumn: false,
-								isJSON: false,
-								key: 'k8s_deployment_name',
-								type: 'tag',
-							},
-							{
-								dataType: DataTypes.String,
-								id: 'k8s_namespace_name--string--tag--false',
-								isColumn: false,
-								isJSON: false,
-								key: 'k8s_namespace_name',
-								type: 'tag',
-							},
-						],
-						having: [],
-						legend: 'desired',
-						limit: null,
-						orderBy: [],
-						queryName: 'B',
-						reduceTo: 'last',
-						spaceAggregation: 'sum',
-						stepInterval: 60,
-						timeAggregation: 'avg',
+						timeAggregation: 'increase',
 					},
 				],
 				queryFormulas: [],
@@ -788,7 +903,7 @@ export const getClusterQueryPayload = (
 			queryType: EQueryType.QUERY_BUILDER,
 		},
 		variables: {},
-		formatForWeb: true,
+		formatForWeb: false,
 		start,
 		end,
 	},
@@ -807,24 +922,24 @@ export const getClusterQueryPayload = (
 							key: 'k8s_statefulset_current_pods',
 							type: 'Gauge',
 						},
-						aggregateOperator: 'max',
+						aggregateOperator: 'latest',
 						dataSource: DataSource.METRICS,
 						disabled: false,
 						expression: 'A',
 						filters: {
 							items: [
 								{
-									id: '3c57b4d1',
+									id: '5f2a55c5',
 									key: {
 										dataType: DataTypes.String,
-										id: 'k8s_cluster_name--string--tag--false',
+										id: 'k8s_namespace_name--string--tag--false',
 										isColumn: false,
 										isJSON: false,
-										key: 'k8s_cluster_name',
+										key: 'k8s_namespace_name',
 										type: 'tag',
 									},
 									op: '=',
-									value: cluster.meta.k8s_cluster_name,
+									value: namespace.namespaceName,
 								},
 							],
 							op: 'AND',
@@ -839,14 +954,6 @@ export const getClusterQueryPayload = (
 								key: 'k8s_statefulset_name',
 								type: 'tag',
 							},
-							{
-								dataType: DataTypes.String,
-								id: 'k8s_namespace_name--string--tag--false',
-								isColumn: false,
-								isJSON: false,
-								key: 'k8s_namespace_name',
-								type: 'tag',
-							},
 						],
 						having: [],
 						legend: 'current',
@@ -854,9 +961,9 @@ export const getClusterQueryPayload = (
 						orderBy: [],
 						queryName: 'A',
 						reduceTo: 'last',
-						spaceAggregation: 'sum',
+						spaceAggregation: 'max',
 						stepInterval: 60,
-						timeAggregation: 'max',
+						timeAggregation: 'latest',
 					},
 					{
 						aggregateAttribute: {
@@ -867,24 +974,24 @@ export const getClusterQueryPayload = (
 							key: 'k8s_statefulset_desired_pods',
 							type: 'Gauge',
 						},
-						aggregateOperator: 'max',
+						aggregateOperator: 'latest',
 						dataSource: DataSource.METRICS,
 						disabled: false,
 						expression: 'B',
 						filters: {
 							items: [
 								{
-									id: '0f49fe64',
+									id: '13bd7a1d',
 									key: {
 										dataType: DataTypes.String,
-										id: 'k8s_cluster_name--string--tag--false',
+										id: 'k8s_namespace_name--string--tag--false',
 										isColumn: false,
 										isJSON: false,
-										key: 'k8s_cluster_name',
+										key: 'k8s_namespace_name',
 										type: 'tag',
 									},
 									op: '=',
-									value: cluster.meta.k8s_cluster_name,
+									value: namespace.namespaceName,
 								},
 							],
 							op: 'AND',
@@ -897,14 +1004,6 @@ export const getClusterQueryPayload = (
 								isColumn: false,
 								isJSON: false,
 								key: 'k8s_statefulset_name',
-								type: 'tag',
-							},
-							{
-								dataType: DataTypes.String,
-								id: 'k8s_namespace_name--string--tag--false',
-								isColumn: false,
-								isJSON: false,
-								key: 'k8s_namespace_name',
 								type: 'tag',
 							},
 						],
@@ -914,69 +1013,9 @@ export const getClusterQueryPayload = (
 						orderBy: [],
 						queryName: 'B',
 						reduceTo: 'last',
-						spaceAggregation: 'sum',
+						spaceAggregation: 'max',
 						stepInterval: 60,
-						timeAggregation: 'max',
-					},
-					{
-						aggregateAttribute: {
-							dataType: DataTypes.Float64,
-							id: 'k8s_statefulset_ready_pods--float64--Gauge--true',
-							isColumn: true,
-							isJSON: false,
-							key: 'k8s_statefulset_ready_pods',
-							type: 'Gauge',
-						},
-						aggregateOperator: 'max',
-						dataSource: DataSource.METRICS,
-						disabled: false,
-						expression: 'C',
-						filters: {
-							items: [
-								{
-									id: '0bebf625',
-									key: {
-										dataType: DataTypes.String,
-										id: 'k8s_cluster_name--string--tag--false',
-										isColumn: false,
-										isJSON: false,
-										key: 'k8s_cluster_name',
-										type: 'tag',
-									},
-									op: '=',
-									value: cluster.meta.k8s_cluster_name,
-								},
-							],
-							op: 'AND',
-						},
-						functions: [],
-						groupBy: [
-							{
-								dataType: DataTypes.String,
-								id: 'k8s_statefulset_name--string--tag--false',
-								isColumn: false,
-								isJSON: false,
-								key: 'k8s_statefulset_name',
-								type: 'tag',
-							},
-							{
-								dataType: DataTypes.String,
-								id: 'k8s_namespace_name--string--tag--false',
-								isColumn: false,
-								isJSON: false,
-								key: 'k8s_namespace_name',
-								type: 'tag',
-							},
-						],
-						having: [],
-						legend: 'ready',
-						limit: null,
-						orderBy: [],
-						queryName: 'C',
-						reduceTo: 'last',
-						spaceAggregation: 'sum',
-						stepInterval: 60,
-						timeAggregation: 'max',
+						timeAggregation: 'latest',
 					},
 					{
 						aggregateAttribute: {
@@ -987,24 +1026,24 @@ export const getClusterQueryPayload = (
 							key: 'k8s_statefulset_updated_pods',
 							type: 'Gauge',
 						},
-						aggregateOperator: 'max',
+						aggregateOperator: 'latest',
 						dataSource: DataSource.METRICS,
 						disabled: false,
-						expression: 'D',
+						expression: 'C',
 						filters: {
 							items: [
 								{
-									id: '1ddacbbe',
+									id: '9d287c73',
 									key: {
 										dataType: DataTypes.String,
-										id: 'k8s_cluster_name--string--tag--false',
+										id: 'k8s_namespace_name--string--tag--false',
 										isColumn: false,
 										isJSON: false,
-										key: 'k8s_cluster_name',
+										key: 'k8s_namespace_name',
 										type: 'tag',
 									},
 									op: '=',
-									value: cluster.meta.k8s_cluster_name,
+									value: namespace.namespaceName,
 								},
 							],
 							op: 'AND',
@@ -1019,24 +1058,16 @@ export const getClusterQueryPayload = (
 								key: 'k8s_statefulset_name',
 								type: 'tag',
 							},
-							{
-								dataType: DataTypes.String,
-								id: 'k8s_namespace_name--string--tag--false',
-								isColumn: false,
-								isJSON: false,
-								key: 'k8s_namespace_name',
-								type: 'tag',
-							},
 						],
 						having: [],
 						legend: 'updated',
 						limit: null,
 						orderBy: [],
-						queryName: 'D',
+						queryName: 'C',
 						reduceTo: 'last',
-						spaceAggregation: 'sum',
+						spaceAggregation: 'max',
 						stepInterval: 60,
-						timeAggregation: 'max',
+						timeAggregation: 'latest',
 					},
 				],
 				queryFormulas: [],
@@ -1074,30 +1105,30 @@ export const getClusterQueryPayload = (
 					{
 						aggregateAttribute: {
 							dataType: DataTypes.Float64,
-							id: 'k8s_daemonset_current_scheduled_nodes--float64--Gauge--true',
+							id: 'k8s_replicaset_desired--float64--Gauge--true',
 							isColumn: true,
 							isJSON: false,
-							key: 'k8s_daemonset_current_scheduled_nodes',
+							key: 'k8s_replicaset_desired',
 							type: 'Gauge',
 						},
-						aggregateOperator: 'avg',
+						aggregateOperator: 'latest',
 						dataSource: DataSource.METRICS,
 						disabled: false,
 						expression: 'A',
 						filters: {
 							items: [
 								{
-									id: 'e0bea554',
+									id: '0c1e655c',
 									key: {
 										dataType: DataTypes.String,
-										id: 'k8s_cluster_name--string--tag--false',
+										id: 'k8s_namespace_name--string--tag--false',
 										isColumn: false,
 										isJSON: false,
-										key: 'k8s_cluster_name',
+										key: 'k8s_namespace_name',
 										type: 'tag',
 									},
 									op: '=',
-									value: cluster.meta.k8s_cluster_name,
+									value: namespace.namespaceName,
 								},
 							],
 							op: 'AND',
@@ -1106,50 +1137,56 @@ export const getClusterQueryPayload = (
 						groupBy: [
 							{
 								dataType: DataTypes.String,
-								id: 'k8s_daemonset_name--string--tag--false',
+								id: 'k8s_replicaset_name--string--tag--false',
 								isColumn: false,
 								isJSON: false,
-								key: 'k8s_daemonset_name',
+								key: 'k8s_replicaset_name',
 								type: 'tag',
 							},
 						],
-						having: [],
-						legend: 'current_nodes',
+						having: [
+							{
+								columnName: 'MAX(k8s_replicaset_desired)',
+								op: '>',
+								value: 0,
+							},
+						],
+						legend: 'desired',
 						limit: null,
 						orderBy: [],
 						queryName: 'A',
 						reduceTo: 'last',
-						spaceAggregation: 'avg',
+						spaceAggregation: 'max',
 						stepInterval: 60,
-						timeAggregation: 'avg',
+						timeAggregation: 'latest',
 					},
 					{
 						aggregateAttribute: {
 							dataType: DataTypes.Float64,
-							id: 'k8s_daemonset_desired_scheduled_nodes--float64--Gauge--true',
+							id: 'k8s_replicaset_available--float64--Gauge--true',
 							isColumn: true,
 							isJSON: false,
-							key: 'k8s_daemonset_desired_scheduled_nodes',
+							key: 'k8s_replicaset_available',
 							type: 'Gauge',
 						},
-						aggregateOperator: 'avg',
+						aggregateOperator: 'latest',
 						dataSource: DataSource.METRICS,
 						disabled: false,
 						expression: 'B',
 						filters: {
 							items: [
 								{
-									id: '741052f7',
+									id: 'b2296bdb',
 									key: {
 										dataType: DataTypes.String,
-										id: 'k8s_cluster_name--string--tag--false',
+										id: 'k8s_namespace_name--string--tag--false',
 										isColumn: false,
 										isJSON: false,
-										key: 'k8s_cluster_name',
+										key: 'k8s_namespace_name',
 										type: 'tag',
 									},
 									op: '=',
-									value: cluster.meta.k8s_cluster_name,
+									value: namespace.namespaceName,
 								},
 							],
 							op: 'AND',
@@ -1158,74 +1195,28 @@ export const getClusterQueryPayload = (
 						groupBy: [
 							{
 								dataType: DataTypes.String,
-								id: 'k8s_daemonset_name--string--tag--false',
+								id: 'k8s_replicaset_name--string--tag--false',
 								isColumn: false,
 								isJSON: false,
-								key: 'k8s_daemonset_name',
+								key: 'k8s_replicaset_name',
 								type: 'tag',
 							},
 						],
-						having: [],
-						legend: 'desired_nodes',
+						having: [
+							{
+								columnName: 'MAX(k8s_replicaset_available)',
+								op: '>',
+								value: 0,
+							},
+						],
+						legend: 'available',
 						limit: null,
 						orderBy: [],
 						queryName: 'B',
 						reduceTo: 'last',
-						spaceAggregation: 'avg',
+						spaceAggregation: 'max',
 						stepInterval: 60,
-						timeAggregation: 'avg',
-					},
-					{
-						aggregateAttribute: {
-							dataType: DataTypes.Float64,
-							id: 'k8s_daemonset_ready_nodes--float64--Gauge--true',
-							isColumn: true,
-							isJSON: false,
-							key: 'k8s_daemonset_ready_nodes',
-							type: 'Gauge',
-						},
-						aggregateOperator: 'avg',
-						dataSource: DataSource.METRICS,
-						disabled: false,
-						expression: 'C',
-						filters: {
-							items: [
-								{
-									id: 'f23759f2',
-									key: {
-										dataType: DataTypes.String,
-										id: 'k8s_cluster_name--string--tag--false',
-										isColumn: false,
-										isJSON: false,
-										key: 'k8s_cluster_name',
-										type: 'tag',
-									},
-									op: '=',
-									value: cluster.meta.k8s_cluster_name,
-								},
-							],
-							op: 'AND',
-						},
-						functions: [],
-						groupBy: [
-							{
-								dataType: DataTypes.String,
-								id: 'k8s_daemonset_name--string--tag--false',
-								isColumn: false,
-								isJSON: false,
-								key: 'k8s_daemonset_name',
-								type: 'tag',
-							},
-						],
-						having: [],
-						legend: 'ready_nodes',
-						limit: null,
-						orderBy: [],
-						queryName: 'C',
-						reduceTo: 'last',
-						spaceAggregation: 'avg',
-						stepInterval: 60,
-						timeAggregation: 'avg',
+						timeAggregation: 'latest',
 					},
 				],
 				queryFormulas: [],
@@ -1263,30 +1254,30 @@ export const getClusterQueryPayload = (
 					{
 						aggregateAttribute: {
 							dataType: DataTypes.Float64,
-							id: 'k8s_job_active_pods--float64--Gauge--true',
+							id: 'k8s_daemonset_desired_scheduled_namespaces--float64--Gauge--true',
 							isColumn: true,
 							isJSON: false,
-							key: 'k8s_job_active_pods',
+							key: 'k8s_daemonset_desired_scheduled_namespaces',
 							type: 'Gauge',
 						},
-						aggregateOperator: 'max',
+						aggregateOperator: 'latest',
 						dataSource: DataSource.METRICS,
 						disabled: false,
 						expression: 'A',
 						filters: {
 							items: [
 								{
-									id: 'fd485d85',
+									id: '2964eb92',
 									key: {
 										dataType: DataTypes.String,
-										id: 'k8s_cluster_name--string--tag--false',
+										id: 'k8s_namespace_name--string--tag--false',
 										isColumn: false,
 										isJSON: false,
-										key: 'k8s_cluster_name',
+										key: 'k8s_namespace_name',
 										type: 'tag',
 									},
 									op: '=',
-									value: cluster.meta.k8s_cluster_name,
+									value: namespace.namespaceName,
 								},
 							],
 							op: 'AND',
@@ -1295,58 +1286,50 @@ export const getClusterQueryPayload = (
 						groupBy: [
 							{
 								dataType: DataTypes.String,
-								id: 'k8s_job_name--string--tag--false',
+								id: 'k8s_daemonset_name--string--tag--false',
 								isColumn: false,
 								isJSON: false,
-								key: 'k8s_job_name',
-								type: 'tag',
-							},
-							{
-								dataType: DataTypes.String,
-								id: 'k8s_namespace_name--string--tag--false',
-								isColumn: false,
-								isJSON: false,
-								key: 'k8s_namespace_name',
+								key: 'k8s_daemonset_name',
 								type: 'tag',
 							},
 						],
 						having: [],
-						legend: 'running',
+						legend: 'desired',
 						limit: null,
 						orderBy: [],
 						queryName: 'A',
 						reduceTo: 'last',
-						spaceAggregation: 'sum',
+						spaceAggregation: 'max',
 						stepInterval: 60,
-						timeAggregation: 'max',
+						timeAggregation: 'latest',
 					},
 					{
 						aggregateAttribute: {
 							dataType: DataTypes.Float64,
-							id: 'k8s_job_successful_pods--float64--Gauge--true',
+							id: 'k8s_daemonset_current_scheduled_namespaces--float64--Gauge--true',
 							isColumn: true,
 							isJSON: false,
-							key: 'k8s_job_successful_pods',
+							key: 'k8s_daemonset_current_scheduled_namespaces',
 							type: 'Gauge',
 						},
-						aggregateOperator: 'max',
+						aggregateOperator: 'latest',
 						dataSource: DataSource.METRICS,
 						disabled: false,
 						expression: 'B',
 						filters: {
 							items: [
 								{
-									id: 'fc1beb81',
+									id: 'cd324eff',
 									key: {
 										dataType: DataTypes.String,
-										id: 'k8s_cluster_name--string--tag--false',
+										id: 'k8s_namespace_name--string--tag--false',
 										isColumn: false,
 										isJSON: false,
-										key: 'k8s_cluster_name',
+										key: 'k8s_namespace_name',
 										type: 'tag',
 									},
 									op: '=',
-									value: cluster.meta.k8s_cluster_name,
+									value: namespace.namespaceName,
 								},
 							],
 							op: 'AND',
@@ -1355,58 +1338,50 @@ export const getClusterQueryPayload = (
 						groupBy: [
 							{
 								dataType: DataTypes.String,
-								id: 'k8s_job_name--string--tag--false',
+								id: 'k8s_daemonset_name--string--tag--false',
 								isColumn: false,
 								isJSON: false,
-								key: 'k8s_job_name',
-								type: 'tag',
-							},
-							{
-								dataType: DataTypes.String,
-								id: 'k8s_namespace_name--string--tag--false',
-								isColumn: false,
-								isJSON: false,
-								key: 'k8s_namespace_name',
+								key: 'k8s_daemonset_name',
 								type: 'tag',
 							},
 						],
 						having: [],
-						legend: 'successful',
+						legend: 'current',
 						limit: null,
 						orderBy: [],
 						queryName: 'B',
 						reduceTo: 'last',
-						spaceAggregation: 'sum',
+						spaceAggregation: 'max',
 						stepInterval: 60,
-						timeAggregation: 'max',
+						timeAggregation: 'latest',
 					},
 					{
 						aggregateAttribute: {
 							dataType: DataTypes.Float64,
-							id: 'k8s_job_failed_pods--float64--Gauge--true',
+							id: 'k8s_daemonset_ready_namespaces--float64--Gauge--true',
 							isColumn: true,
 							isJSON: false,
-							key: 'k8s_job_failed_pods',
+							key: 'k8s_daemonset_ready_namespaces',
 							type: 'Gauge',
 						},
-						aggregateOperator: 'max',
+						aggregateOperator: 'latest',
 						dataSource: DataSource.METRICS,
 						disabled: false,
 						expression: 'C',
 						filters: {
 							items: [
 								{
-									id: '97773348',
+									id: '0416fa6f',
 									key: {
 										dataType: DataTypes.String,
-										id: 'k8s_cluster_name--string--tag--false',
+										id: 'k8s_namespace_name--string--tag--false',
 										isColumn: false,
 										isJSON: false,
-										key: 'k8s_cluster_name',
+										key: 'k8s_namespace_name',
 										type: 'tag',
 									},
 									op: '=',
-									value: cluster.meta.k8s_cluster_name,
+									value: namespace.namespaceName,
 								},
 							],
 							op: 'AND',
@@ -1415,58 +1390,50 @@ export const getClusterQueryPayload = (
 						groupBy: [
 							{
 								dataType: DataTypes.String,
-								id: 'k8s_job_name--string--tag--false',
+								id: 'k8s_daemonset_name--string--tag--false',
 								isColumn: false,
 								isJSON: false,
-								key: 'k8s_job_name',
-								type: 'tag',
-							},
-							{
-								dataType: DataTypes.String,
-								id: 'k8s_namespace_name--string--tag--false',
-								isColumn: false,
-								isJSON: false,
-								key: 'k8s_namespace_name',
+								key: 'k8s_daemonset_name',
 								type: 'tag',
 							},
 						],
 						having: [],
-						legend: 'failed',
+						legend: 'ready',
 						limit: null,
 						orderBy: [],
 						queryName: 'C',
 						reduceTo: 'last',
-						spaceAggregation: 'sum',
+						spaceAggregation: 'max',
 						stepInterval: 60,
-						timeAggregation: 'max',
+						timeAggregation: 'latest',
 					},
 					{
 						aggregateAttribute: {
 							dataType: DataTypes.Float64,
-							id: 'k8s_job_desired_successful_pods--float64--Gauge--true',
+							id: 'k8s_daemonset_misscheduled_namespaces--float64--Gauge--true',
 							isColumn: true,
 							isJSON: false,
-							key: 'k8s_job_desired_successful_pods',
+							key: 'k8s_daemonset_misscheduled_namespaces',
 							type: 'Gauge',
 						},
-						aggregateOperator: 'max',
+						aggregateOperator: 'latest',
 						dataSource: DataSource.METRICS,
 						disabled: false,
 						expression: 'D',
 						filters: {
 							items: [
 								{
-									id: '5911618b',
+									id: 'c0a126d3',
 									key: {
 										dataType: DataTypes.String,
-										id: 'k8s_cluster_name--string--tag--false',
+										id: 'k8s_namespace_name--string--tag--false',
 										isColumn: false,
 										isJSON: false,
-										key: 'k8s_cluster_name',
+										key: 'k8s_namespace_name',
 										type: 'tag',
 									},
 									op: '=',
-									value: cluster.meta.k8s_cluster_name,
+									value: namespace.namespaceName,
 								},
 							],
 							op: 'AND',
@@ -1475,33 +1442,169 @@ export const getClusterQueryPayload = (
 						groupBy: [
 							{
 								dataType: DataTypes.String,
-								id: 'k8s_job_name--string--tag--false',
+								id: 'k8s_daemonset_name--string--tag--false',
 								isColumn: false,
 								isJSON: false,
-								key: 'k8s_job_name',
-								type: 'tag',
-							},
-							{
-								dataType: DataTypes.String,
-								id: 'k8s_namespace_name--string--tag--false',
-								isColumn: false,
-								isJSON: false,
-								key: 'k8s_namespace_name',
+								key: 'k8s_daemonset_name',
 								type: 'tag',
 							},
 						],
 						having: [],
-						legend: 'desired successful',
+						legend: 'misscheduled',
 						limit: null,
 						orderBy: [],
 						queryName: 'D',
 						reduceTo: 'last',
-						spaceAggregation: 'sum',
+						spaceAggregation: 'max',
 						stepInterval: 60,
-						timeAggregation: 'max',
+						timeAggregation: 'latest',
 					},
 				],
 				queryFormulas: [],
+			},
+			clickhouse_sql: [
+				{
+					disabled: false,
+					legend: '',
+					name: 'A',
+					query: '',
+				},
+			],
+			id: v4(),
+			promql: [
+				{
+					disabled: false,
+					legend: '',
+					name: 'A',
+					query: '',
+				},
+			],
+			queryType: EQueryType.QUERY_BUILDER,
+		},
+		variables: {},
+		formatForWeb: true,
+		start,
+		end,
+	},
+	{
+		selectedTime: 'GLOBAL_TIME',
+		graphType: PANEL_TYPES.TABLE,
+		query: {
+			builder: {
+				queryData: [
+					{
+						aggregateAttribute: {
+							dataType: DataTypes.Float64,
+							id: 'k8s_deployment_desired--float64--Gauge--true',
+							isColumn: true,
+							isJSON: false,
+							key: 'k8s_deployment_desired',
+							type: 'Gauge',
+						},
+						aggregateOperator: 'latest',
+						dataSource: DataSource.METRICS,
+						disabled: false,
+						expression: 'A',
+						filters: {
+							items: [
+								{
+									id: '9bc659c1',
+									key: {
+										dataType: DataTypes.String,
+										id: 'k8s_namespace_name--string--tag--false',
+										isColumn: false,
+										isJSON: false,
+										key: 'k8s_namespace_name',
+										type: 'tag',
+									},
+									op: '=',
+									value: namespace.namespaceName,
+								},
+							],
+							op: 'AND',
+						},
+						functions: [],
+						groupBy: [
+							{
+								dataType: DataTypes.String,
+								id: 'k8s_deployment_name--string--tag--false',
+								isColumn: false,
+								isJSON: false,
+								key: 'k8s_deployment_name',
+								type: 'tag',
+							},
+						],
+						having: [],
+						legend: 'desired',
+						limit: null,
+						orderBy: [],
+						queryName: 'A',
+						reduceTo: 'last',
+						spaceAggregation: 'max',
+						stepInterval: 60,
+						timeAggregation: 'latest',
+					},
+					{
+						aggregateAttribute: {
+							dataType: DataTypes.Float64,
+							id: 'k8s_deployment_available--float64--Gauge--true',
+							isColumn: true,
+							isJSON: false,
+							key: 'k8s_deployment_available',
+							type: 'Gauge',
+						},
+						aggregateOperator: 'avg',
+						dataSource: DataSource.METRICS,
+						disabled: false,
+						expression: 'B',
+						filters: {
+							items: [
+								{
+									id: 'e1696631',
+									key: {
+										dataType: DataTypes.String,
+										id: 'k8s_namespace_name--string--tag--false',
+										isColumn: false,
+										isJSON: false,
+										key: 'k8s_namespace_name',
+										type: 'tag',
+									},
+									op: '=',
+									value: namespace.namespaceName,
+								},
+							],
+							op: 'AND',
+						},
+						functions: [],
+						groupBy: [
+							{
+								dataType: DataTypes.String,
+								id: 'k8s_deployment_name--string--tag--false',
+								isColumn: false,
+								isJSON: false,
+								key: 'k8s_deployment_name',
+								type: 'tag',
+							},
+						],
+						having: [],
+						legend: 'available',
+						limit: null,
+						orderBy: [],
+						queryName: 'B',
+						reduceTo: 'last',
+						spaceAggregation: 'max',
+						stepInterval: 60,
+						timeAggregation: 'avg',
+					},
+				],
+				queryFormulas: [
+					{
+						disabled: false,
+						expression: 'A/B',
+						legend: 'util %',
+						queryName: 'F1',
+					},
+				],
 			},
 			clickhouse_sql: [
 				{
